PowerSystems.Branch Formulations

Valid DeviceModels for subtypes of Branch include the following:

Valid DeviceModel	Device Type	Formulation

DeviceModel(PhaseShiftingTransformer, PhaseAngleControl)	PhaseShiftingTransformer	PhaseAngleControl

DeviceModel(HVDCLine, HVDCP2PDispatch)	HVDCLine	HVDCP2PDispatch

DeviceModel(HVDCLine, HVDCP2PLossless)	HVDCLine	HVDCP2PLossless

DeviceModel(VSCDCLine, HVDCP2PLossless)	VSCDCLine	HVDCP2PLossless

DeviceModel(HVDCLine, HVDCP2PUnbounded)	HVDCLine	HVDCP2PUnbounded

DeviceModel(VSCDCLine, HVDCP2PUnbounded)	VSCDCLine	HVDCP2PUnbounded

DeviceModel(Line, StaticBranch)	Line	StaticBranch

DeviceModel(MonitoredLine, StaticBranch)	MonitoredLine	StaticBranch

DeviceModel(PhaseShiftingTransformer, StaticBranch)	PhaseShiftingTransformer	StaticBranch

DeviceModel(TapTransformer, StaticBranch)	TapTransformer	StaticBranch

DeviceModel(Transformer2W, StaticBranch)	Transformer2W	StaticBranch

DeviceModel(Line, StaticBranchBounds)	Line	StaticBranchBounds

DeviceModel(MonitoredLine, StaticBranchBounds)	MonitoredLine	StaticBranchBounds

DeviceModel(PhaseShiftingTransformer, StaticBranchBounds)	PhaseShiftingTransformer	StaticBranchBounds

DeviceModel(TapTransformer, StaticBranchBounds)	TapTransformer	StaticBranchBounds

DeviceModel(Transformer2W, StaticBranchBounds)	Transformer2W	StaticBranchBounds

DeviceModel(Line, StaticBranchUnbounded)	Line	StaticBranchUnbounded

DeviceModel(MonitoredLine, StaticBranchUnbounded)	MonitoredLine	StaticBranchUnbounded

DeviceModel(PhaseShiftingTransformer, StaticBranchUnbounded)	PhaseShiftingTransformer	StaticBranchUnbounded

DeviceModel(TapTransformer, StaticBranchUnbounded)	TapTransformer	StaticBranchUnbounded

DeviceModel(Transformer2W, StaticBranchUnbounded)	Transformer2W	StaticBranchUnbounded

StaticBranch

PowerSimulations.StaticBranch — Type

Branch type to add unbounded flow variables and use flow constraints

StaticBranchBounds

PowerSimulations.StaticBranchBounds — Type

Branch type to add bounded flow variables and use flow constraints

StaticBranchUnbounded

PowerSimulations.StaticBranchUnbounded — Type

Branch type to avoid flow constraints

HVDCP2PLossless

PowerSimulations.HVDCP2PLossless — Type

Branch type to represent lossless power flow on DC lines

HVDCP2PDispatch

PowerSimulations.HVDCP2PDispatch — Type

Branch type to represent lossy power flow on DC lines

HVDCP2PUnbounded

PowerSimulations.HVDCP2PUnbounded — Type

Branch type to avoid flow constraints