PowerSystems.HydroGen Formulations

Valid DeviceModels for subtypes of HydroGen include the following:

Valid DeviceModel	Device Type	Formulation

DeviceModel(HydroDispatch, FixedOutput)	HydroDispatch	FixedOutput

DeviceModel(HydroEnergyReservoir, FixedOutput)	HydroEnergyReservoir	FixedOutput

DeviceModel(HydroPumpedStorage, FixedOutput)	HydroPumpedStorage	FixedOutput

DeviceModel(HydroDispatch, HydroDispatchRunOfRiver)	HydroDispatch	HydroDispatchRunOfRiver

DeviceModel(HydroEnergyReservoir, HydroDispatchRunOfRiver)	HydroEnergyReservoir	HydroDispatchRunOfRiver

DeviceModel(HydroPumpedStorage, HydroDispatchRunOfRiver)	HydroPumpedStorage	HydroDispatchRunOfRiver

DeviceModel(HydroDispatch, HydroDispatchPumpedStorage)	HydroDispatch	HydroDispatchPumpedStorage

DeviceModel(HydroEnergyReservoir, HydroDispatchPumpedStorage)	HydroEnergyReservoir	HydroDispatchPumpedStorage

DeviceModel(HydroPumpedStorage, HydroDispatchPumpedStorage)	HydroPumpedStorage	HydroDispatchPumpedStorage

DeviceModel(HydroDispatch, HydroDispatchReservoirBudget)	HydroDispatch	HydroDispatchReservoirBudget

DeviceModel(HydroEnergyReservoir, HydroDispatchReservoirBudget)	HydroEnergyReservoir	HydroDispatchReservoirBudget

DeviceModel(HydroPumpedStorage, HydroDispatchReservoirBudget)	HydroPumpedStorage	HydroDispatchReservoirBudget

DeviceModel(HydroDispatch, HydroDispatchReservoirStorage)	HydroDispatch	HydroDispatchReservoirStorage

DeviceModel(HydroEnergyReservoir, HydroDispatchReservoirStorage)	HydroEnergyReservoir	HydroDispatchReservoirStorage

DeviceModel(HydroPumpedStorage, HydroDispatchReservoirStorage)	HydroPumpedStorage	HydroDispatchReservoirStorage

DeviceModel(HydroDispatch, HydroCommitmentReservoirBudget)	HydroDispatch	HydroCommitmentReservoirBudget

DeviceModel(HydroEnergyReservoir, HydroCommitmentReservoirBudget)	HydroEnergyReservoir	HydroCommitmentReservoirBudget

DeviceModel(HydroPumpedStorage, HydroCommitmentReservoirBudget)	HydroPumpedStorage	HydroCommitmentReservoirBudget

DeviceModel(HydroEnergyReservoir, HydroCommitmentReservoirStorage)	HydroEnergyReservoir	HydroCommitmentReservoirStorage

DeviceModel(HydroDispatch, HydroCommitmentRunOfRiver)	HydroDispatch	HydroCommitmentRunOfRiver

DeviceModel(HydroEnergyReservoir, HydroCommitmentRunOfRiver)	HydroEnergyReservoir	HydroCommitmentRunOfRiver

DeviceModel(HydroPumpedStorage, HydroCommitmentRunOfRiver)	HydroPumpedStorage	HydroCommitmentRunOfRiver

HydroDispatchRunOfRiver

PowerSimulations.HydroDispatchRunOfRiver — Type

Formulation type to add injection variables constrained by a maximum injection time series for HydroGen

HydroDispatchRunOfRiver()

defined at /home/runner/work/PowerSimulations.jl/PowerSimulations.jl/src/core/formulations.jl:90.

Variables:

ActivePowerVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_active_power(device)

ReactivePowerVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_reactive_power(device)

Static Parameters:

$Pg^\text{min}$ = PowerSystems.get_active_power_limits(device).min

$Qg^\text{min}$ = PowerSystems.get_reactive_power_limits(device).min

$Qg^\text{max}$ = PowerSystems.get_reactive_power_limits(device).max

Time Series Parameters:

Parameter	Default Time Series Name

ActivePowerTimeSeriesParameter	max_active_power

ReactivePowerTimeSeriesParameter	max_active_power

Objective:

Creates an objective function term based on the VariableCost Options where the quantity term is defined as $Pg$.

Expressions:

Adds $Pg$ and $Qg$ terms to the respective active and reactive power balance expressions created by the selected Network Formulations

Constraints:

\[\begin{aligned} & Pg^\text{min} \le Pg_t \le ActivePowerTimeSeriesParameter_t \\ & Qg^\text{min} \le Qg_t \le Qg^\text{max} \end{aligned}\]

HydroDispatchPumpedStorage

PowerSimulations.HydroDispatchPumpedStorage — Type

Formulation type to constrain energy production from pumped storage with a representation of the energy storage capacity of upper and lower reservoirs and water inflow time series of upper reservoir and outflow time series of lower reservoir for HydroPumpedStorage

HydroDispatchPumpedStorage()

defined at /home/runner/work/PowerSimulations.jl/PowerSimulations.jl/src/core/formulations.jl:105.

Variables:

ActivePowerInVariable:

Bounds: [0.0, ]

Default initial value: -1 * PowerSystems.get_active_power(device)

ActivePowerOutVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_active_power(device)

EnergyVariableUp:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_initial_storage(device).up

EnergyVariableDown:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_initial_storage(device).down

WaterSpillageVariable:

Bounds: [0.0, ]

Default initial value: 0.0

EnergyOutput:

Bounds: [0.0, ]

Default initial value: 0.0

ReservationVariable:

only included if DeviceModel(HydroPumpedStorage, HydroDispatchPumpedStorage; attributes = Dict(reservation => true))

Bounds: {0, 1}

Default initial value: 1

Static Parameters:

$Pg^\text{out, max}$ = map(x -> x.max - x.min, PowerSystems.get_active_power_limits(device))

$Pg^\text{in, max}$ = map(x -> x.max - x.min, PowerSystems.get_active_power_limits_pump(device))

$Eg^\text{up, max}$ = PowerSystems.get_storage_capacity(device).up

$Eg^\text{down, max}$ = PowerSystems.get_storage_capacity(device).down

Time Series Parameters:

Parameter	Default Time Series Name

PowerSimulations.OutflowTimeSeriesParameter	outflow

PowerSimulations.InflowTimeSeriesParameter	inflow

Objective:

Creates an objective function term based on the VariableCost Options where the quantity term is defined as $Pg^{out}$.

Expressions:

Adds $Pg$ term(s) to the active power balance expression(s) created by the selected Network Formulations

Constraints:

\[\begin{aligned} & E^{up}_{t+1} = E^{up}_t + (InflowTimeSeriesParameter_t - S_t - Pg^{out}_t + Pg^{in}_t) \cdot \Delta T \\ & E^{down}_{t+1} = E^{down}_t + (S_t - OutflowTimeSeriesParameter_t + Pg^{out}_t - Pg^{in}_t) \cdot \Delta T \\ & Pg^{in}_t - r * Pg^\text{in, max} \le Pg^\text{in, max} \\ & Pg^{out}_t + r * Pg^\text{out, max} \le Pg^\text{out, max} \\ & E^{up}_t \le E^{up, max} & E^{down}_t \le E^{down, max} \end{aligned}\]

HydroDispatchReservoirBudget

PowerSimulations.HydroDispatchReservoirBudget — Type

Formulation type to add injection variables constrained by total energy production budget defined with a time series for HydroGen

HydroDispatchReservoirBudget()

defined at /home/runner/work/PowerSimulations.jl/PowerSimulations.jl/src/core/formulations.jl:95.

Variables:

ActivePowerVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_active_power(device)

ReactivePowerVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_reactive_power(device)

Auxillary Variables:

EnergyOutput - TODO

Static Parameters:

$Pg^\text{min}$ = PowerSystems.get_active_power_limits(device).min

$Pg^\text{max}$ = PowerSystems.get_active_power_limits(device).max

$Qg^\text{min}$ = PowerSystems.get_reactive_power_limits(device).min

$Qg^\text{max}$ = PowerSystems.get_reactive_power_limits(device).max

Time Series Parameters:

Parameter	Default Time Series Name

EnergyBudgetTimeSeriesParameter	hydro_budget

Objective:

Creates an objective function term based on the VariableCost Options where the quantity term is defined as $Pg$.

Expressions:

Adds $Pg$ and $Qg$ terms to the respective active and reactive power balance expressions created by the selected Network Formulations

Constraints:

\[\begin{aligned} & Pg^\text{min} \le Pg_t \le Pg^\text{max} \\ & Qg^\text{min} \le Qg_t \le Qg^\text{max} \\ & \sum_{t = 1}^N(Pg_t) \cdot \Delta T \le \sum_{t = 1}^N(EnergyBudgetTimeSeriesParameter_t) \cdot \Delta T \end{aligned}\]

HydroDispatchReservoirStorage

PowerSimulations.HydroDispatchReservoirStorage — Type

Formulation type to constrain hydropower production with a representation of the energy storage capacity and water inflow time series of a reservoir for HydroGen

HydroDispatchReservoirStorage()

defined at /home/runner/work/PowerSimulations.jl/PowerSimulations.jl/src/core/formulations.jl:100.

Variables:

ActivePowerVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_active_power(device)

ReactivePowerVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_reactive_power(device)

EnergyVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_initial_storage(device)

WaterSpillageVariable:

Bounds: [0.0, ]

Default initial value: 0.0

EnergyShortageVariable:

Bounds: [ , 0.0]

Default initial value: 0.0

EnergySurplusVariable:

Bounds: [0.0, ]

Default initial value: 0.0

Auxillary Variables:

EnergyOutput

Static Parameters:

$Pg^\text{min}$ = PowerSystems.get_active_power_limits(device).min

$Pg^\text{max}$ = PowerSystems.get_active_power_limits(device).max

$Qg^\text{min}$ = PowerSystems.get_reactive_power_limits(device).min

$Qg^\text{max}$ = PowerSystems.get_reactive_power_limits(device).max

$Eg^\text{max}$ = PowerSystems.get_storage_capacity(device)

Time Series Parameters:

Parameter	Default Time Series Name

PowerSimulations.InflowTimeSeriesParameter	inflow

EnergyTargetTimeSeriesParameter	storage_target

Objective:

Creates an objective function term based on the VariableCost Options where the quantity term is defined as $Pg$. TODO: add slack terms

Expressions:

Adds $Pg$ and $Qg$ terms to the respective active and reactive power balance expressions created by the selected Network Formulations

Constraints:

\[\begin{aligned} & E_{t+1} = E_t + (InflowTimeSeriesParameter_t - S_t - Pg_t) \cdot \Delta T \\ & E_t - E^{surplus}_t + E^{shortage}_t = EnergyTargetTimeSeriesParameter_t \\ & Pg^\text{min} \le Pg_t \le Pg^\text{max} \\ & Qg^\text{min} \le Qg_t \le Qg^\text{max} \end{aligned}\]

HydroCommitmentReservoirBudget

PowerSimulations.HydroCommitmentReservoirBudget — Type

Formulation type to add commitment and injection variables constrained by total energy production budget defined with a time series for HydroGen

HydroCommitmentReservoirBudget()

defined at /home/runner/work/PowerSimulations.jl/PowerSimulations.jl/src/core/formulations.jl:115.

Variables:

ActivePowerVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_active_power(device)

ReactivePowerVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_reactive_power(device)

OnVariable:

Bounds: {0, 1}

Default initial value: PowerSystems.get_status(device)

Auxillary Variables:

EnergyOutput - TODO

Static Parameters:

$Pg^\text{min}$ = PowerSystems.get_active_power_limits(device).min

$Pg^\text{max}$ = PowerSystems.get_active_power_limits(device).max

$Qg^\text{min}$ = PowerSystems.get_reactive_power_limits(device).min

$Qg^\text{max}$ = PowerSystems.get_reactive_power_limits(device).max

Time Series Parameters:

Parameter	Default Time Series Name

EnergyBudgetTimeSeriesParameter	hydro_budget

Objective:

Creates an objective function term based on the VariableCost Options where the quantity term is defined as $Pg$.

Expressions:

Adds $Pg$ and $Qg$ terms to the respective active and reactive power balance expressions created by the selected Network Formulations

Constraints:

\[\begin{aligned} & \sum_{t = 1}^N P_t \cdot \Delta T \le E^\text{budget} \\ & Pg_t - u_t Pg^\text{max} \le 0 \\ & Pg_t - u_t Pg^\text{min} \ge 0 \\ & Qg_t - u_t Qg^\text{max} \le 0 \\ & Qg_t - u_t Qg^\text{min} \ge 0 \end{aligned}\]

HydroCommitmentReservoirStorage

PowerSimulations.HydroCommitmentReservoirStorage — Type

Formulation type to constrain hydropower production with unit commitment variables and a representation of the energy storage capacity and water inflow time series of a reservoir for HydroGen

HydroCommitmentReservoirStorage()

defined at /home/runner/work/PowerSimulations.jl/PowerSimulations.jl/src/core/formulations.jl:120.

Variables:

ActivePowerVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_active_power(device)

ReactivePowerVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_reactive_power(device)

EnergyVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_initial_storage(device)

WaterSpillageVariable:

Bounds: [0.0, ]

Default initial value: 0.0

EnergyShortageVariable:

Bounds: [ , 0.0]

Default initial value: 0.0

EnergySurplusVariable:

Bounds: [0.0, ]

Default initial value: 0.0

OnVariable:

Bounds: {0, 1}

Default initial value: PowerSystems.get_status(device)

Auxillary Variables:

EnergyOutput

Static Parameters:

$Pg^\text{min}$ = PowerSystems.get_active_power_limits(device).min

$Pg^\text{max}$ = PowerSystems.get_active_power_limits(device).max

$Qg^\text{min}$ = PowerSystems.get_reactive_power_limits(device).min

$Qg^\text{max}$ = PowerSystems.get_reactive_power_limits(device).max

$Eg^\text{max}$ = PowerSystems.get_storage_capacity(device)

Time Series Parameters:

Parameter	Default Time Series Name

PowerSimulations.InflowTimeSeriesParameter	inflow

EnergyTargetTimeSeriesParameter	storage_target

Objective:

Creates an objective function term based on the VariableCost Options where the quantity term is defined as $Pg$, and objective function terms for StorageManagementCost.

Expressions:

Adds $Pg$ and $Qg$ terms to the respective active and reactive power balance expressions created by the selected Network Formulations

Constraints:

\[\begin{aligned} & E_{t+1} = E_t + (InflowTimeSeriesParameter_t - S_t - Pg_t) \cdot \Delta T \\ & E_t - E^{surplus}_t + E^{shortage}_t = EnergyTargetTimeSeriesParameter_t \\ & Pg_t - u_t Pg^\text{max} \le 0 \\ & Pg_t - u_t Pg^\text{min} \ge 0 \\ & Qg_t - u_t Qg^\text{max} \le 0 \\ & Qg_t - u_t Qg^\text{min} \ge 0 \end{aligned}\]

HydroCommitmentRunOfRiver

PowerSimulations.HydroCommitmentRunOfRiver — Type

Formulation type to add commitment and injection variables constrained by a maximum injection time series for HydroGen

HydroCommitmentRunOfRiver()

defined at /home/runner/work/PowerSimulations.jl/PowerSimulations.jl/src/core/formulations.jl:110.

Variables:

ActivePowerVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_active_power(device)

ReactivePowerVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_reactive_power(device)

OnVariable:

Bounds: {0, 1}

Default initial value: PowerSystems.get_status(device)

Static Parameters:

$Pg^\text{min}$ = PowerSystems.get_active_power_limits(device).min

$Qg^\text{min}$ = PowerSystems.get_reactive_power_limits(device).min

$Qg^\text{max}$ = PowerSystems.get_reactive_power_limits(device).max

Time Series Parameters:

Parameter	Default Time Series Name

ActivePowerTimeSeriesParameter	max_active_power

ReactivePowerTimeSeriesParameter	max_active_power

Objective:

Creates an objective function term based on the VariableCost Options where the quantity term is defined as $Pg$.

Expressions:

Adds $Pg$ and $Qg$ terms to the respective active and reactive power balance expressions created by the selected Network Formulations

Constraints:

\[\begin{aligned} & Pg_t \le Pg^\text{max}\\ & Pg_t - u_t Pg^\text{max} \le 0 \\ & Pg_t - u_t Pg^\text{min} \ge 0 \\ & Qg_t - u_t Qg^\text{max} \le 0 \\ & Qg_t - u_t Qg^\text{min} \ge 0 \end{aligned}\]