PowerSystems.ElectricLoad Formulations

Valid DeviceModels for subtypes of ElectricLoad include the following:

Valid DeviceModel	Device Type	Formulation

DeviceModel(InterruptibleLoad, DispatchablePowerLoad)	InterruptibleLoad	DispatchablePowerLoad

DeviceModel(PowerLoad, DispatchablePowerLoad)	PowerLoad	DispatchablePowerLoad

DeviceModel(PowerSystems.ExponentialLoad, DispatchablePowerLoad)	PowerSystems.ExponentialLoad	DispatchablePowerLoad

DeviceModel(StandardLoad, DispatchablePowerLoad)	StandardLoad	DispatchablePowerLoad

DeviceModel(InterruptibleLoad, InterruptiblePowerLoad)	InterruptibleLoad	InterruptiblePowerLoad

DeviceModel(PowerLoad, InterruptiblePowerLoad)	PowerLoad	InterruptiblePowerLoad

DeviceModel(PowerSystems.ExponentialLoad, InterruptiblePowerLoad)	PowerSystems.ExponentialLoad	InterruptiblePowerLoad

DeviceModel(StandardLoad, InterruptiblePowerLoad)	StandardLoad	InterruptiblePowerLoad

DeviceModel(InterruptibleLoad, StaticPowerLoad)	InterruptibleLoad	StaticPowerLoad

DeviceModel(FixedAdmittance, StaticPowerLoad)	FixedAdmittance	StaticPowerLoad

DeviceModel(PowerLoad, StaticPowerLoad)	PowerLoad	StaticPowerLoad

DeviceModel(PowerSystems.ExponentialLoad, StaticPowerLoad)	PowerSystems.ExponentialLoad	StaticPowerLoad

DeviceModel(StandardLoad, StaticPowerLoad)	StandardLoad	StaticPowerLoad

StaticPowerLoad

PowerSimulations.StaticPowerLoad — Type

Formulation type to add a time series paraemter for non-dispatchable ElectricLoad withdrawls to power balance constraints

StaticPowerLoad()

defined at /home/runner/work/PowerSimulations.jl/PowerSimulations.jl/src/core/formulations.jl:63.

Variables:

No variables are created

Time Series Parameters:

Parameter	Default Time Series Name

ActivePowerTimeSeriesParameter	max_active_power

ReactivePowerTimeSeriesParameter	max_active_power

Expressions:

Subtracts the parameters listed above from the respective active and reactive power balance expressions created by the selected Network Formulations

Constraints:

No constraints are created

InterruptiblePowerLoad

PowerSimulations.InterruptiblePowerLoad — Type

Formulation type to enable (binary) load interruptions

InterruptiblePowerLoad()

defined at /home/runner/work/PowerSimulations.jl/PowerSimulations.jl/src/core/formulations.jl:68.

Variables:

ActivePowerVariable:

Bounds: [0.0, ]

Default initial value: 0.0

ReactivePowerVariable:

Bounds: [0.0, ]

Default initial value: 0.0

OnVariable:

Bounds: {0,1}

Default initial value: 1

Time Series Parameters:

Parameter	Default Time Series Name

ActivePowerTimeSeriesParameter	max_active_power

ReactivePowerTimeSeriesParameter	max_active_power

Objective:

Creates an objective function term based on the VariableCost Options where the quantity term is defined as $Pg$.

Expressions:

Adds $Pg$ and $Qg$ terms and to the respective active and reactive power balance expressions created by the selected Network Formulations

Subtracts the time series parameters listed above terms from the respective active and reactive power balance expressions created by the selected Network Formulations

Constraints:

$Pg$ and $Qg$ represent the "unserved" active and reactive power loads

\[\begin{aligned} & Pg_t \le ActivePowerTimeSeriesParameter_t\\ & Pg_t - u_t ActivePowerTimeSeriesParameter_t \le 0 \\ & Qg_t \le ReactivePowerTimeSeriesParameter_t\\ & Qg_t - u_t ReactivePowerTimeSeriesParameter_t\le 0 \end{aligned}\]

DispatchablePowerLoad

PowerSimulations.DispatchablePowerLoad — Type

Formulation type to enable (continuous) load interruption dispatch

DispatchablePowerLoad()

defined at /home/runner/work/PowerSimulations.jl/PowerSimulations.jl/src/core/formulations.jl:73.

Variables:

ActivePowerVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_active_power(device)

ReactivePowerVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_reactive_power(device)

Time Series Parameters:

Parameter	Default Time Series Name

ActivePowerTimeSeriesParameter	max_active_power

ReactivePowerTimeSeriesParameter	max_active_power

Objective:

Creates an objective function term based on the VariableCost Options where the quantity term is defined as $Pg$.

Expressions:

Adds $Pg$ and $Qg$ terms and to the respective active and reactive power balance expressions created by the selected Network Formulations

Subtracts the time series parameters listed above terms from the respective active and reactive power balance expressions created by the selected Network Formulations

Constraints:

$Pg$ and $Qg$ represent the "unserved" active and reactive power loads

\[\begin{aligned} & Pg_t \le ActivePowerTimeSeriesParameter_t\\ & Qg_t \le ReactivePowerTimeSeriesParameter_t\\ \end{aligned}\]