PowerSystems.Storage Formulations

Valid DeviceModels for subtypes of Storage include the following:

Valid DeviceModel	Device Type	Formulation

DeviceModel(BatteryEMS, BatteryAncillaryServices)	BatteryEMS	BatteryAncillaryServices

DeviceModel(GenericBattery, BatteryAncillaryServices)	GenericBattery	BatteryAncillaryServices

DeviceModel(BatteryEMS, BookKeeping)	BatteryEMS	BookKeeping

DeviceModel(GenericBattery, BookKeeping)	GenericBattery	BookKeeping

DeviceModel(BatteryEMS, EnergyTarget)	BatteryEMS	EnergyTarget

DeviceModel(GenericBattery, EnergyTarget)	GenericBattery	EnergyTarget

BookKeeping

PowerSimulations.BookKeeping — Type

Formulation type to add basic storage formulation. With attributes=Dict("reservation"=>true) the formulation is augmented with abinary variable to prevent simultanious charging and discharging

BookKeeping()

defined at /home/runner/work/PowerSimulations.jl/PowerSimulations.jl/src/core/formulations.jl:149.

Variables:

ActivePowerInVariable:

Bounds: [0.0, ]

Default initial value: -1 * PowerSystems.get_active_power(device)

ActivePowerOutVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_active_power(device)

ReactivePowerVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_reactive_power(device)

EnergyVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_initial_storage(device)

ReservationVariable:

only included if DeviceModel(HydroPumpedStorage, HydroDispatchPumpedStorage; attributes = Dict(reservation => true))

Bounds: {0, 1}

Default initial value: 1

Static Parameters:

$Pg^\text{min}$ = PowerSystems.get_active_power_limits(device).min

$Qg^\text{min}$ = PowerSystems.get_reactive_power_limits(device).min

$Qg^\text{max}$ = PowerSystems.get_reactive_power_limits(device).max

$E^\text{max}$ = PowerSystems.get_storage_capacity(device)

Objective:

Creates an objective function term based on the VariableCost Options where the quantity term is defined as $Pg_t$.

Expressions:

Adds $Pg$ and $Qg$ terms to the respective active and reactive power balance expressions created by the selected Network Formulations

Constraints:

\[\begin{aligned} & E_{t+1} = E_t + (Pg^{in}_t - Pg^{out}_t) \cdot \Delta T \\ & Pg^{in}_t - r * Pg^\text{in, max} \le Pg^\text{in, max} \\ & Pg^{out}_t + r * Pg^\text{out, max} \le Pg^\text{out, max} \\ & Qg^\text{min} \le Qg_t \le Qg^\text{max} \\ & E_t \le E^\text{max} \end{aligned}\]

EnergyTarget

PowerSimulations.EnergyTarget — Type

Formulation type to add storage formulation that respects end of horizon energy state of charge target. With attributes=Dict("reservation"=>true) the formulation is augmented with abinary variable to prevent simultanious charging and discharging

EnergyTarget()

defined at /home/runner/work/PowerSimulations.jl/PowerSimulations.jl/src/core/formulations.jl:159.

Variables:

ActivePowerInVariable:

Bounds: [0.0, ]

Default initial value: -1 * PowerSystems.get_active_power(device)

ActivePowerOutVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_active_power(device)

ReactivePowerVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_reactive_power(device)

EnergyVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_initial_storage(device)

EnergyShortageVariable:

Bounds: [ , 0.0]

Default initial value: 0.0

EnergySurplusVariable:

Bounds: [0.0, ]

Default initial value: 0.0

ReservationVariable:

only included if DeviceModel(HydroPumpedStorage, HydroDispatchPumpedStorage; attributes = Dict(reservation => true))

Bounds: {0, 1}

Default initial value: 1

Static Parameters:

$Pg^\text{min}$ = PowerSystems.get_active_power_limits(device).min

$Qg^\text{min}$ = PowerSystems.get_reactive_power_limits(device).min

$Qg^\text{max}$ = PowerSystems.get_reactive_power_limits(device).max

$E^\text{max}$ = PowerSystems.get_storage_capacity(device)

Time Series Parameters:

Parameter	Default Time Series Name

EnergyTargetTimeSeriesParameter	storage_target

Objective:

Creates an objective function term based on the VariableCost Options where the quantity term is defined as $Pg_t$, and objective function terms for StorageManagementCost.

Expressions:

Adds $Pg$ and $Qg$ terms to the respective active and reactive power balance expressions created by the selected Network Formulations

Constraints:

\[\begin{aligned} & E_{t+1} = E_t + (Pg^{in}_t - Pg^{out}_t) \cdot \Delta T \\ & E_t - E^{surplus}_t + E^{shortage}_t = EnergyTargetTimeSeriesParameter_t \\ & Pg^{in}_t - r * Pg^\text{in, max} \le Pg^\text{in, max} \\ & Pg^{out}_t + r * Pg^\text{out, max} \le Pg^\text{out, max} \\ & Qg^\text{min} \le Qg_t \le Qg^\text{max}\\ & E_t \le E^\text{max} \end{aligned}\]

BatteryAncillaryServices

PowerSimulations.BatteryAncillaryServices — Type

Formulation type to add storage formulation than can provide ancillary services. With attributes=Dict("reservation"=>true) the formulation is augmented with abinary variable to prevent simultanious charging and discharging

BatteryAncillaryServices()

defined at /home/runner/work/PowerSimulations.jl/PowerSimulations.jl/src/core/formulations.jl:154.

Variables:

ActivePowerInVariable:

Bounds: [0.0, ]

Default initial value: -1 * PowerSystems.get_active_power(device)

ActivePowerOutVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_active_power(device)

ReactivePowerVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_reactive_power(device)

EnergyVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_initial_storage(device)

ReservationVariable:

only included if DeviceModel(HydroPumpedStorage, HydroDispatchPumpedStorage; attributes = Dict(reservation => true))

Bounds: {0, 1}

Default initial value: 1

Static Parameters:

$Pg^\text{min}$ = PowerSystems.get_active_power_limits(device).min

$Qg^\text{min}$ = PowerSystems.get_reactive_power_limits(device).min

$Qg^\text{max}$ = PowerSystems.get_reactive_power_limits(device).max

$E^\text{max}$ = PowerSystems.get_storage_capacity(device)

Time Series Parameters:

Parameter	Default Time Series Name

EnergyTargetTimeSeriesParameter	storage_target

Objective:

Creates an objective function term based on the VariableCost Options where the quantity term is defined as $Pg_t$, and objective function terms for StorageManagementCost.

Expressions:

Adds $Pg$ and $Qg$ terms to the respective active and reactive power balance expressions created by the selected Network Formulations

Constraints:

\[\begin{aligned} & E_{t+1} = E_t + (Pg^{in}_t - Pg^{out}_t) \cdot \Delta T \\ & E_t - E^{surplus}_t + E^{shortage}_t = EnergyTargetTimeSeriesParameter_t \\ & Pg^{in}_t - r * Pg^\text{in, max} \le Pg^\text{in, max} \\ & Pg^{out}_t + r * Pg^\text{out, max} \le Pg^\text{out, max} \\ & Qg^\text{min} \le Qg_t \le Qg^\text{max}\\ & E_t \le E^\text{max} \end{aligned}\]