Device Models

List of structures and methods for Device models

PowerSimulations.DeviceModel — Type

Establishes the model for a particular device specified by type. Uses the keyword argument feedforward to enable passing values between operation model at simulation time

Arguments

-::Type{D}: Power System Device Type -::Type{B}: Abstract Device Formulation

Accepted Key Words

feedforward::Array{<:AbstractAffectFeedforward} : use to pass parameters between models

use_slacks::Bool : Add slacks to the device model

duals::Vector{DataType} : use to pass constraint type to calculate the duals

time_series_names::Dict{Type{<:TimeSeriesParameter}, String} : use to specify time series names associated to the device

attributes::Dict{String, Any} : use to specify attributes to the device

Example

thermal_gens = DeviceModel(ThermalStandard, ThermalBasicUnitCommitment),

DeviceModel(
    ,
    ;
    feedforwards,
    use_slacks,
    duals,
    time_series_names,
    attributes
)

defined at /home/runner/work/PowerSimulations.jl/PowerSimulations.jl/src/core/device_model.jl:47.

Formulations

Refer to the Formulations Page for each Abstract Device Formulation.

Decision Models

PowerSimulations.DecisionModel — Type

DecisionModel(::Type{M},
template::ProblemTemplate,
sys::PSY.System,
jump_model::Union{Nothing, JuMP.Model}=nothing;
kwargs...) where {M<:DecisionProblem,
                  T<:PM.AbstractPowerFormulation}

This builds the optimization problem of type M with the specific system and template.

Arguments

::Type{M} where M<:DecisionProblem: The abstract operation model type

template::ProblemTemplate: The model reference made up of transmission, devices, branches, and services.

sys::PSY.System: the system created using Power Systems

jump_model::Union{Nothing, JuMP.Model}: Enables passing a custom JuMP model. Use with care

Output

model::DecisionModel: The operation model containing the model type, built JuMP model, Power Systems system.

Example

template = ProblemTemplate(CopperPlatePowerModel, devices, branches, services)
OpModel = DecisionModel(MockOperationProblem, template, system)

Accepted Key Words

optimizer: The optimizer that will be used in the optimization model.

horizon::Int: Manually specify the length of the forecast Horizon

warm_start::Bool: True will use the current operation point in the system to initialize variable values. False initializes all variables to zero. Default is true

system_to_file::Bool:: True to create a copy of the system used in the model. Default true.

export_pwl_vars::Bool: True to export all the pwl intermediate variables. It can slow down significantly the solve time. Default is false.

allow_fails::Bool: True to allow the simulation to continue even if the optimization step fails. Use with care, default to false.

optimizer_solve_log_print::Bool: True to print the optimizer solve log. Default is false.

direct_mode_optimizer::Bool True to use the solver in direct mode. Creates a JuMP.direct_model. Default is false.

initial_time::Dates.DateTime: Initial Time for the model solve

time_series_cache_size::Int: Size in bytes to cache for each time array. Default is 1 MiB. Set to 0 to disable.

DecisionModel(, template, sys)
DecisionModel(, template, sys, jump_model; kwargs...)

defined at /home/runner/work/PowerSimulations.jl/PowerSimulations.jl/src/operation/decision_model.jl:176.

DecisionModel(template, sys)
DecisionModel(template, sys, jump_model; kwargs...)

defined at /home/runner/work/PowerSimulations.jl/PowerSimulations.jl/src/operation/decision_model.jl:186.

DecisionModel(directory, optimizer; jump_model, system)

defined at /home/runner/work/PowerSimulations.jl/PowerSimulations.jl/src/operation/decision_model.jl:211.

PowerSimulations.build! — Method

build!(
    model::DecisionModel;
    output_dir,
    recorders,
    console_level,
    file_level,
    disable_timer_outputs
)

Build the Decision Model based on the specified DecisionProblem.

Arguments

model::DecisionModel{<:DecisionProblem}: DecisionModel object

output_dir::String: Output directory for results

Accepted Key Words

recorders::Vector{Symbol} = []: recorder names to register

console_level = Logging.Error:

file_level = Logging.Info:

disable_timer_outputs = false : Enable/Disable timing outputs

build!(
    model;
    output_dir,
    recorders,
    console_level,
    file_level,
    disable_timer_outputs
)

defined at /home/runner/work/PowerSimulations.jl/PowerSimulations.jl/src/operation/decision_model.jl:297.

PowerSimulations.solve! — Method

solve!(
    model::DecisionModel;
    export_problem_results,
    console_level,
    file_level,
    disable_timer_outputs,
    serialize,
    kwargs...
) -> RunStatus

Default solve method for models that conform to the requirements of DecisionModel{<: DecisionProblem}.

This will call build! on the model if it is not already built. It will forward all keyword arguments to that function.

Arguments

model::OperationModel = model: operation model

Accepted Key Words

export_problem_results::Bool = false: If true, export ProblemResults DataFrames to CSV files.

console_level = Logging.Error:

file_level = Logging.Info:

disable_timer_outputs = false : Enable/Disable timing outputs

serialize::Bool = true: If true, serialize the model to a file to allow re-execution later.

Examples

results = solve!(OpModel)
results = solve!(OpModel, export_problem_results = true)

solve!(
    model;
    export_problem_results,
    console_level,
    file_level,
    disable_timer_outputs,
    serialize,
    kwargs...
)

defined at /home/runner/work/PowerSimulations.jl/PowerSimulations.jl/src/operation/decision_model.jl:390.