PowerSimulations

PowerSimulations

API documentation

PowerSimulations

Index

Exported

Internal

Index

PowerSimulations.TimeStatusChange

PowerSimulations.UpdateRef

Base.show

PowerSimulations._container_spec

PowerSimulations._get_data_for_rocc

PowerSimulations._get_data_for_tdc

PowerSimulations._internal_device_constructor!

PowerSimulations._internal_device_constructor!

PowerSimulations._internal_device_constructor!

PowerSimulations._internal_device_constructor!

PowerSimulations._internal_device_constructor!

PowerSimulations._internal_device_constructor!

PowerSimulations._pwl_cost

PowerSimulations._pwlgencost_linear

PowerSimulations._pwlgencost_sos

PowerSimulations._pwlparamcheck

PowerSimulations._retrieve_forecasts

PowerSimulations.activepower_constraints

PowerSimulations.activepower_constraints!

PowerSimulations.activepower_constraints!

PowerSimulations.activepower_constraints!

PowerSimulations.activepower_variables!

PowerSimulations.add_to_cost

PowerSimulations.add_variable

PowerSimulations.build_nip_expr_model

PowerSimulations.build_nip_model

PowerSimulations.commitment_constraints!

PowerSimulations.commitment_variables!

PowerSimulations.constraint_power_balance_ni

PowerSimulations.constraint_power_balance_ni

PowerSimulations.constraint_power_balance_ni

PowerSimulations.constraint_power_balance_ni_expr

PowerSimulations.constraint_power_balance_ni_expr

PowerSimulations.constraint_power_balance_ni_expr

PowerSimulations.device_commitment

PowerSimulations.device_duration_look_ahead

PowerSimulations.device_duration_param

PowerSimulations.device_duration_retrospective

PowerSimulations.device_linear_rateofchange

PowerSimulations.device_mixedinteger_rateofchange

PowerSimulations.device_range

PowerSimulations.device_semicontinuousrange

PowerSimulations.device_timeseries_lb

PowerSimulations.device_timeseries_param_lb

PowerSimulations.device_timeseries_param_ub

PowerSimulations.device_timeseries_ub

PowerSimulations.device_timeseries_ub_bigM

PowerSimulations.device_timeseries_ub_bin

PowerSimulations.energy_balance

PowerSimulations.get_all_constraint_index

PowerSimulations.get_all_var_index

PowerSimulations.parameter_update!

PowerSimulations.post_nip

PowerSimulations.post_nip_expr

PowerSimulations.powermodels_network!

PowerSimulations.powermodels_network!

PowerSimulations.ps_cost

PowerSimulations.ps_cost

PowerSimulations.ps_cost

PowerSimulations.ps_cost

PowerSimulations.ramp_constraints!

PowerSimulations.range_ff

PowerSimulations.rating_constraint

PowerSimulations.reactive_power_constraints

PowerSimulations.reactivepower_constraints

PowerSimulations.reactivepower_constraints!

PowerSimulations.reactivepower_constraints!

PowerSimulations.reactivepower_variables!

PowerSimulations.remove_chars

PowerSimulations.replace_chars

PowerSimulations.reserve_device_semicontinuousrange

PowerSimulations.run_sim_model!

PowerSimulations.semicontinuousrange_ff

PowerSimulations.set_variable_bounds

PowerSimulations.sort_data

PowerSimulations.status_init

PowerSimulations.ub_ff

PowerSimulations.variable_active_net_injection

PowerSimulations.variable_net_injection

PowerSimulations.variable_reactive_net_injection

PowerSimulations.variable_reactive_net_injection

PowerSimulations.write_model_results

PowerSimulations.write_op_model

Exported

PowerSimulations.TimeStatusChange — Type.

Tracks the last time status of a device changed in a simulation

PowerSimulations.get_all_constraint_index — Method.

"Each Tuple corresponds to (conname, internalindex, moi_index)

PowerSimulations.get_all_var_index — Method.

"Each Tuple corresponds to (conname, internalindex, moi_index)

PowerSimulations.run_sim_model! — Method.

Runs Simulations

PowerSimulations.sort_data — Method.

sort_data(results::OperationModelResults)

This function takes in struct OperationModelResults, sorts the generators in each variable, and outputs the sorted results. The generic function sorts the generators alphabetically.

kwargs: 'Variables' to choose which variables to be sorted.

each variable has a kwarg, so that a specific order can be generated, such that when plotted, the first generator is on the bottom. if a list of generator names has fewer generators than the variable, only the generators on the list will be outputted.

#Examples

example 1: sortedresults = sortdata(res)

sortedresults.variables[PRenewableDispatch] will be in the order

[:WindBusA :WindBusB :WindBusC] (alphabetical)

example 2: myorder = [:WindBusC :WindBusB :WindBusA] sortedresults = sortdata(res; PRenewableDispatch = my_order)

sortedresults.variables[PRenewableDispatch] will be in the order

[:WindBusC :WindBusB :WindBusA] (my_order)

example 3: myorder = [:WindBusC :WindBusA] sortedresults = sortdata(res; PRenewableDispatch = my_order)

sortedresults.variables[PRenewableDispatch] will be in the order

[:WindBusC :WindBusA] (my_order)

example 4: myvariableorder = [:PThermalStandard :ONThermalStandard] sortedresults = sortdata(res; Variables = myvariableorder)

sorted_results.variables

Dict{Symbol,DataFrames.DataFrame} with 2 entries:
:P_ThermalStandard => 24×5 DataFrames.DataFrame…
:ON_ThermalStandard => 24×5 DataFrames.DataFrame…

note that only the generators included in 'my_order' will be in the

results, and consequently, only these will be plotted. This can be a nice feature for variables with more than 5 generators.

PowerSimulations.write_model_results — Method.

Exports Operational Model Results to a path

PowerSimulations.write_op_model — Method.

Exports the OpModel JuMP object in MathOptFormat

Internal

PowerSimulations.UpdateRef — Type.

Reference for parameters update when present

Base.show — Method.

Base.show(io::IO, ::MIME"text/plain", op_model::OperationModel)

This function goes through the fields in OperationModel and then in ModelReference, if the field contains a Device model dictionary, it calls organizedevicemodel() & prints the data by field, key, value. If the field is not a Device model dictionary, and a value exists for that field it prints the value.

PowerSimulations._container_spec — Method.

Returns the correct container spec for the selected type of JuMP Model

PowerSimulations._get_data_for_rocc — Method.

This function gets the data for the generators

PowerSimulations._get_data_for_tdc — Method.

If the fraction of hours that a generator has a duration constraint is less than the fraction of hours that a single time_step represents then it is not binding.

PowerSimulations._internal_device_constructor! — Method.

This function creates the model for a full themal dispatch formulation depending on combination of devices, deviceformulation and systemformulation

PowerSimulations._internal_device_constructor! — Method.

This function creates the model for a full themal dispatch formulation depending on combination of devices, deviceformulation and systemformulation

PowerSimulations._internal_device_constructor! — Method.

This function creates the model for a full themal dispatch formulation depending on combination of devices, deviceformulation and systemformulation

PowerSimulations._internal_device_constructor! — Method.

This function creates the model for a full themal dispatch formulation depending on combination of devices, deviceformulation and systemformulation

PowerSimulations._internal_device_constructor! — Method.

This function creates the model for a full themal dispatch formulation depending on combination of devices, deviceformulation and systemformulation

PowerSimulations._internal_device_constructor! — Method.

This function creates the model for a full themal dispatch formulation depending on combination of devices, deviceformulation and systemformulation

PowerSimulations._pwl_cost — Method.

_gen_cost(cost_)

Returns JuMP expression for a piecewise linear cost function depending on the data compatibility.

Returns gen_cost

Arguments

canonical_model::CanonicalModel : the canonical model built in PowerSimulations

variable::JuMP.Containers.DenseAxisArray{JV} : variable array

cost_component::PSY.VariableCost{NTuple{2, Float64}} : container for quadratic and linear factors

PowerSimulations._pwlgencost_linear — Method.

_pwlgencost_sos(canonical_model::CanonicalModel,
            variable::JuMP.Containers.DenseAxisArray{JV},
            cost_component::PSY.VariableCost{NTuple{2, Float64}}) where {JV <: JuMP.AbstractVariableRef}

Returns piecewise cost expression using linear implementation for canonical model.

Equations

0 <= pwl_var[i] <= (cost_component[2][i] - cost_component[2][i-1])

variable = sum(pwl_var[i])

gen_cost = sum(pwl_var[i]*cost_component[1][i]/cost_component[2][i])

LaTeX

$0 <= pwl_i <= (c_{2, i} - c_{2, i-1})$

$variable = (sum_{i\in I} pwl_i)$

$gen_cost = (sum_{i\in I} pwl_i) c_{1, i}/c_{2, i}$

Returns gen_cost

Arguments

canonical_model::CanonicalModel : the canonical model built in PowerSimulations

variable::JuMP.Containers.DenseAxisArray{JV} : variable array

cost_component::PSY.VariableCost{NTuple{2, Float64}} : container for quadratic and linear factors

PowerSimulations._pwlgencost_sos — Method.

_pwlgencost_sos(canonical_model::CanonicalModel,
            variable::JuMP.Containers.DenseAxisArray{JV},
            cost_component::PSY.VariableCost{NTuple{2, Float64}}) where {JV <: JuMP.AbstractVariableRef}

Returns piecewise cost expression using SOS Type-2 implementation for canonical model.

Equations

variable = sum(sos_var[i]*cost_component[2][i])

gen_cost = sum(sos_var[i]*cost_component[1][i])

LaTeX

$variable = (sum_{i\in I} c_{2, i} sos_i)$

$gen_cost = (sum_{i\in I} c_{1, i} sos_i)$

Returns gen_cost

Arguments

canonical_model::CanonicalModel : the canonical model built in PowerSimulations

variable::JuMP.Containers.DenseAxisArray{JV} : variable array

cost_component::PSY.VariableCost{NTuple{2, Float64}} : container for quadratic and linear factors

PowerSimulations._pwlparamcheck — Method.

_pwlparamcheck(cost_)

Returns True/False depending on compatibility of the cost data with the linear implementation method

Returns flag

Arguments

cost_::PSY.VariableCost{NTuple{2, Float64}} : container for quadratic and linear factors

PowerSimulations._retrieve_forecasts — Method.

Only used when building the model

PowerSimulations.activepower_constraints! — Method.

This function adds the active power limits of generators when there are no CommitmentVariables

PowerSimulations.activepower_constraints! — Method.

This function adds the active power limits of generators when there are CommitmentVariables

PowerSimulations.activepower_constraints! — Method.

This function adds the active power limits of generators when there are no CommitmentVariables

PowerSimulations.activepower_constraints — Method.

This function works only if the the ParamL <= PSY.getmaxactivepower(g)

PowerSimulations.activepower_variables! — Method.

This function add the variables for power generation output to the model

PowerSimulations.add_to_cost — Method.

add_to_cost(canonical_model::CanonicalModel,
                 devices::D,
                 var_name::Symbol,
                 cost_symbol::Symbol,
                 sign::Float64 = 1.0) where {D<:IS.FlattenIteratorWrapper{<:PSY.Device}}

Adds cost expression for each device using appropriate call to ps_cost.

Expression

for d in devices

                              variable[PSY.get_name(d), :],
                              getfield(PSY.get_op_cost(d), cost_symbol),
                              dt,
                              sign) ```

canonicalmodel.costfunction += cost_expression ```

LaTeX

$COST = \sum_{d\in D} cost_d$

Arguments

canonical_model::CanonicalModel : the canonical model built in PowerSimulations

devices::D : set of devices

var_name::Symbol : name of variable

cost_symbol::Symbol : symbol associated with costx

PowerSimulations.add_variable — Method.

add_variable(canonical_model::CanonicalModel,
                  devices::D,
                  var_name::Symbol,
                  binary::Bool,
                  expression::Symbol,
                  sign::Float64)

Adds a variable to the optimization model and to the affine expressions contained in the canonical model according to the specified sign. Based on the inputs, the variable can be specified as binary.

Bounds

lb_value_function <= varstart[name, t] <= ub_value_function

If binary = true:

varstart[name, t] in {0,1}

LaTeX

$lb \ge x^{device}_t \le ub \forall t$

$x^{device}_t \in {0,1} \forall t iff \text{binary = true}$

Arguments

canonical_model::CanonicalModel : the canonical model built in PowerSimulations

devices : Vector or Iterator with the devices

var_name::Symbol : Base Name for the variable

binary::Bool : Select if the variable is binary

expression::Symbol : Expression name stored in canonical_model.expressions to add the variable

sign::Float64 : sign of the addition of the variable to the expression. Default Value is 1.0

Accepted Keyword Arguments

ubvaluefunction : Provides the function over device to obtain the value for a upper_bound

lbvaluefunction : Provides the function over device to obtain the value for a lower_bound. If the variable is meant to be positive define lb = x -> 0.0

initialvaluefunction : Provides the function over device to obtain the warm start value

PowerSimulations.build_nip_expr_model — Method.

PowerSimulations.build_nip_model — Method.

PowerSimulations.commitment_constraints! — Method.

This function adds the Commitment Status constraint when there are CommitmentVariables

PowerSimulations.commitment_variables! — Method.

This function add the variables for power generation commitment to the model

PowerSimulations.constraint_power_balance_ni — Method.

PowerSimulations.constraint_power_balance_ni — Method.

PowerSimulations.constraint_power_balance_ni — Method.

active power only models ignore reactive power flows

PowerSimulations.constraint_power_balance_ni_expr — Method.

PowerSimulations.constraint_power_balance_ni_expr — Method.

PowerSimulations.constraint_power_balance_ni_expr — Method.

PowerSimulations.device_commitment — Method.

device_commitment(canonical_model::CanonicalModel,
                    initial_conditions::Vector{InitialCondition},
                    cons_name::Symbol,
                    var_names::Tuple{Symbol, Symbol, Symbol})

Constructs multi-timestep constraint from initial conditions and binary variable tuple.

Constraints

varstart + varstop <= 1.0

If t = 1:

varon[name, 1] == ic.value + varstart[name, 1] - varstop[name, 1]

where ic in initial_condtions.

If t > 1:

varon[name, t] == varon[name, t-1] + varstart[name, t] - varstop[name, t]

LaTeX

$x^{on}_t + x^{off}_t \leq 1.0 \forall t$

$x^{on}_1 = x^{on}_{init} + x^{start}_1 - x^{stop}_1, \text{ for } t = 1$

$x^{on}_t = x^{on}_{t-1} + x^{start}_t - x^{stop}_t, \forall t \geq 2$

Arguments

canonical_model::CanonicalModel : the canonical model built in PowerSimulations

initial_conditions::Vector{InitialCondition} : for time zero 'varon'

cons_name::Symbol : name of the constraint

var_names::Tuple{Symbol, Symbol, Symbol} : the names of the variables

: var_names[1] : varstart

: var_names[2] : varstop

: var_names[3] : varon

PowerSimulations.device_duration_look_ahead — Method.

device_duration_look_ahead(canonical_model::CanonicalModel,
                            duration_data::Vector{UpDown},
                            initial_duration::Matrix{InitialCondition},
                            cons_name::Symbol,
                            var_names::Tuple{Symbol, Symbol, Symbol})

This formulation of the duration constraints looks ahead in the time frame of the model.

LaTeX

Minimum up-time constraint:

If $t \leq d_{min}^{up}$

$d_{min}^{down}x_t^{stop} - \sum_{i=t-d_{min}^{up} + 1}^t x_i^{on} - x_{init}^{up} \leq 0$

for i in the set of time steps. Otherwise:

$d_{min}^{down}x_t^{stop} - \sum_{i=t-d_{min}^{up} + 1}^t x_i^{on} \leq 0$

for i in the set of time steps.

Minimum down-time constraint:

If $t \leq d_{min}^{down}$

$d_{min}^{up}x_t^{start} - \sum_{i=t-d_{min^{down} + 1}^t (1 - x_i^{on}) - x_{init}^{down} \leq 0$

for i in the set of time steps. Otherwise:

$d_{min}^{up}x_t^{start} - \sum_{i=t-d_{min^{down} + 1}^t (1 - x_i^{on}) \leq 0$

for i in the set of time steps.

Arguments

canonical_model::CanonicalModel : the canonical model built in PowerSimulations

duration_data::Vector{UpDown} : gives how many time steps variable needs to be up or down

initial_duration::Matrix{InitialCondition} : gives initial conditions for up (column 1) and down (column 2)

cons_name::Symbol : name of the constraint

var_names::Tuple{Symbol, Symbol, Symbol}) : names of the variables

: var_names[1] : varon

: var_names[2] : varstart

: var_names[3] : varstop

PowerSimulations.device_duration_param — Method.

device_duration_param(canonical_model::CanonicalModel,
                         duration_data::Vector{UpDown},
                         initial_duration_on::Vector{InitialCondition},
                         initial_duration_off::Vector{InitialCondition},
                         cons_name::Symbol,
                         var_names::Tuple{Symbol, Symbol, Symbol})

This formulation of the duration constraints considers parameters.

LaTeX

Minimum up-time constraint:

If $t \leq d_{min}^{up}$

$d_{min}^{down}x_t^{stop} - \sum_{i=t-d_{min}^{up} + 1}^t x_i^{on} - x_{init}^{up} \leq 0$

for i in the set of time steps. Otherwise:

$\sum_{i=t-d_{min}^{up} + 1}^t x_i^{start} - x_t^{on} \leq 0$

for i in the set of time steps.

Minimum down-time constraint:

If $t \leq d_{min}^{down}$

$d_{min}^{up}x_t^{start} - \sum_{i=t-d_{min^{down} + 1}^t (1 - x_i^{on}) - x_{init}^{down} \leq 0$

for i in the set of time steps. Otherwise:

$\sum_{i=t-d_{min}^{down} + 1}^t x_i^{stop} + x_t^{on} \leq 1$

for i in the set of time steps.

Arguments

canonical_model::CanonicalModel : the canonical model built in PowerSimulations

duration_data::Vector{UpDown} : gives how many time steps variable needs to be up or down

initialdurationon::Vector{InitialCondition} : gives initial number of time steps variable is up

initialdurationoff::Vector{InitialCondition} : gives initial number of time steps variable is down

cons_name::Symbol : name of the constraint

var_names::Tuple{Symbol, Symbol, Symbol}) : names of the variables

: var_names[1] : varon

: var_names[2] : varstart

: var_names[3] : varstop

PowerSimulations.device_duration_retrospective — Method.

device_duration_retrospective(canonical_model::CanonicalModel,
                                    duration_data::Vector{UpDown},
                                    initial_duration::Matrix{InitialCondition},
                                    cons_name::Symbol,
                                    var_names::Tuple{Symbol, Symbol, Symbol})

This formulation of the duration constraints adds over the start times looking backwards.

LaTeX

Minimum up-time constraint:

If $t \leq d_{min}^{up} - d_{init}^{up}$ and $d_{init}^{up} > 0$

$1 + \sum_{i=t-d_{min}^{up} + 1}^t x_i^{start} - x_t^{on} \leq 0$

for i in the set of time steps. Otherwise:

$\sum_{i=t-d_{min}^{up} + 1}^t x_i^{start} - x_t^{on} \leq 0$

for i in the set of time steps.

Minimum down-time constraint:

If $t \leq d_{min}^{down} - d_{init}^{down}$ and $d_{init}^{down} > 0$

$1 + \sum_{i=t-d_{min}^{down} + 1}^t x_i^{stop} + x_t^{on} \leq 1$

for i in the set of time steps. Otherwise:

$\sum_{i=t-d_{min}^{down} + 1}^t x_i^{stop} + x_t^{on} \leq 1$

for i in the set of time steps.

Arguments

canonical_model::CanonicalModel : the canonical model built in PowerSimulations

duration_data::Vector{UpDown} : gives how many time steps variable needs to be up or down

initial_duration::Matrix{InitialCondition} : gives initial conditions for up (column 1) and down (column 2)

cons_name::Symbol : name of the constraint

var_names::Tuple{Symbol, Symbol, Symbol}) : names of the variables

: var_names[1] : varon

: var_names[2] : varstart

: var_names[3] : varstop

PowerSimulations.device_linear_rateofchange — Method.

device_linear_rateofchange(canonical_model::CanonicalModel,
                                rate_data::Tuple{Vector{String}, Vector{UpDown}},
                                initial_conditions::Vector{InitialCondition},
                                cons_name::Symbol,
                                var_name::Symbol)

Constructs allowed rate-of-change constraints from variables, initial condtions, and rate data.

Constraints

If t = 1:

variable[name, 1] - initial_conditions[ix].value <= rate_data[1][ix].up

initial_conditions[ix].value - variable[name, 1] <= rate_data[1][ix].down

If t > 1:

variable[name, t] - variable[name, t-1] <= rate_data[1][ix].up

variable[name, t-1] - variable[name, t] <= rate_data[1][ix].down

LaTeX

$r^{down} \leq x_1 - x_{init} \leq r^{up}, \text{ for } t = 1$

$r^{down} \leq x_t - x_{t-1} \leq r^{up}, \forall t \geq 2$

Arguments

canonical_model::CanonicalModel : the canonical model built in PowerSimulations

rate_data::Tuple{Vector{String}, Vector{UpDown}} : gives name (1) and max ramp up/down rates (2)

initial_conditions::Vector{InitialCondition} : for time zero 'variable'

cons_name::Symbol : name of the constraint

var_name::Tuple{Symbol, Symbol, Symbol} : the name of the variable

PowerSimulations.device_mixedinteger_rateofchange — Method.

device_mixedinteger_rateofchange(canonical_model::CanonicalModel,
                                      rate_data::Tuple{Vector{String}, Vector{UpDown}, Vector{MinMax}},
                                      initial_conditions::Vector{InitialCondition},
                                      cons_name::Symbol,
                                      var_names::Tuple{Symbol, Symbol, Symbol})

Constructs allowed rate-of-change constraints from variables, initial condtions, start/stop status, and rate data

Equations

If t = 1:

variable[name, 1] - initial_conditions[ix].value <= rate_data[1][ix].up + rate_data[2][ix].max*varstart[name, 1]

initial_conditions[ix].value - variable[name, 1] <= rate_data[1][ix].down + rate_data[2][ix].min*varstop[name, 1]

If t > 1:

variable[name, t] - variable[name, t-1] <= rate_data[1][ix].up + rate_data[2][ix].max*varstart[name, t]

variable[name, t-1] - variable[name, t] <= rate_data[1][ix].down + rate_data[2][ix].min*varstop[name, t]

LaTeX

$r^{down} + r^{min} x^{stop}_1 \leq x_1 - x_{init} \leq r^{up} + r^{max} x^{start}_1, \text{ for } t = 1$

$r^{down} + r^{min} x^{stop}_t \leq x_t - x_{t-1} \leq r^{up} + r^{max} x^{start}_t, \forall t \geq 2$

Arguments

canonical_model::CanonicalModel : the canonical model built in PowerSimulations

rate_data::Tuple{Vector{String}, Vector{UpDown}, Vector{MinMax}} : (1) gives name (2) gives min/max ramp rates (3) gives min/max for 'variable'

initial_conditions::Vector{InitialCondition} : for time zero 'variable'

cons_name::Symbol : name of the constraint

var_names::Tuple{Symbol, Symbol, Symbol} : the names of the variables

: var_names[1] : 'variable'

: var_names[2] : 'varstart'

: var_names[3] : 'varstop'

PowerSimulations.device_range — Method.

device_range(canonical_model::CanonicalModel,
                    range_data::Vector{NamedMinMax},
                    cons_name::Symbol,
                    var_name::Symbol)

Constructs min/max range constraint from device variable.

Constraints

If min and max within an epsilon width:

variable[r[1], t] == r[2].max

Otherwise:

r[2].min <= variable[r[1], t] <= r[2].max

where r in range_data.

LaTeX

$x = r^{max}, \text{ for } |r^{max} - r^{min}| < \varepsilon$

$r^{min} \leq x \leq r^{max}, \text{ otherwise }$

Arguments

canonical_model::CanonicalModel : the canonical model built in PowerSimulations

range_data::Vector{NamedMinMax} : contains name of device (1) and its min/max (2)

cons_name::Symbol : name of the constraint

var_name::Symbol : the name of the continuous variable

PowerSimulations.device_semicontinuousrange — Method.

device_semicontinuousrange(canonical_model::CanonicalModel,
                                scrange_data::Vector{NamedMinMax},
                                cons_name::Symbol,
                                var_name::Symbol,
                                binvar_name::Symbol)

Constructs min/max range constraint from device variable and on/off decision variable.

Constraints

If device min = 0:

varcts[r[1], t] <= r[2].max*varbin[r[1], t])

varcts[r[1], t] >= 0.0

Otherwise:

varcts[r[1], t] <= r[2].max*varbin[r[1], t]

varcts[r[1], t] >= r[2].min*varbin[r[1], t]

where r in range_data.

LaTeX

$0 \leq x^{cts} \leq r^{max} x^{bin}, \text{ for } r^{min} = 0$

$r^{min} x^{bin} \leq x^{cts} \leq r^{max} x^{bin}, \text{ otherwise }$

Arguments

canonical_model::CanonicalModel : the canonical model built in PowerSimulations

scrange_data::Vector{NamedMinMax} : contains name of device (1) and its min/max (2)

cons_name::Symbol : name of the constraint

var_name::Symbol : the name of the continuous variable

binvar_name::Symbol : the name of the binary variable

PowerSimulations.device_timeseries_lb — Method.

device_timeseries_lb(canonical_model::CanonicalModel,
                 ts_data::Tuple{Vector{String}, Vector{Float64}, Vector{Vector{Float64}}},
                 cons_name::Symbol,
                 var_name::Symbol)

Constructs lower bound for given variable subject to time series data and a multiplier.

Constraint

ts_data[2][ix]*ts_data[3][ix][t] <= variable[name, t]

LaTeX

$r^{val} r_t \leq x_t, \forall t$

where (ix, name) in enumerate(ts_data[1]).

Arguments

canonical_model::CanonicalModel : the canonical model built in PowerSimulations

ts_data::Tuple{Vector{String}, Vector{Float64}, Vector{Vector{Float64}}} : timeseries data name (1), multiplier (2) and values (3)

cons_name::Symbol : name of the constraint

var_name::Symbol : the name of the variable

PowerSimulations.device_timeseries_param_lb — Method.

device_timeseries_param_lb(canonical_model::CanonicalModel,
                                ts_data::Tuple{Vector{String}, Vector{Float64}, Vector{Vector{Float64}}},
                                cons_name::Symbol,
                                param_reference::UpdateRef,
                                var_name::Symbol)

Constructs lower bound for given variable using a parameter. The constraint is built with a time series data vector and a multiplier

Constraint

ts_data[2][ix] * param[name, t] <= variable[name, t]

LaTeX

$r^{val} x^{param}_t \leq x^{var}_t, \forall t$

Arguments

canonical_model::CanonicalModel : the canonical model built in PowerSimulations

ts_data::Tuple{Vector{String}, Vector{Vector{Float64}}} : timeseries data name (1) and values (2)

cons_name::Symbol : name of the constraint

param_reference::UpdateRef : UpdateRef to access the parameter

var_name::Symbol : the name of the variable

PowerSimulations.device_timeseries_param_ub — Method.

device_timeseries_param_ub(canonical_model::CanonicalModel,
                                ts_data::Tuple{Vector{String}, Vector{Float64}, Vector{Vector{Float64}}},
                                cons_name::Symbol,
                                param_reference::UpdateRef,
                                var_name::Symbol)

Constructs upper bound for given variable using a parameter. The constraint is built with a time series data vector and a multiplier

Constraint

variable[name, t] <= ts_data[2][ix]*param[name, t]

LaTeX

$x^{var}_t \leq r^{val} x^{param}_t, \forall t$

Arguments

canonical_model::CanonicalModel : the canonical model built in PowerSimulations

ts_data::Tuple{Vector{String}, Vector{Float64}, Vector{Vector{Float64}}} : timeseries data name (1), multiplier (2) and values (3)

cons_name::Symbol : name of the constraint

param_reference::UpdateRef : UpdateRef to access the parameter

var_name::Symbol : the name of the variable

PowerSimulations.device_timeseries_ub — Method.

device_timeseries_ub(canonical_model::CanonicalModel,
                 ts_data::Tuple{Vector{String}, Vector{Float64}, Vector{Vector{Float64}}},
                 cons_name::Symbol,
                 var_name::Symbol)

Constructs upper bound for given variable and time series data and a multiplier.

Constraint

variable[name, t] <= ts_data[2][ix]*ts_data[3][ix][t]

LaTeX

$x_t \leq r^{val} r_t, \forall t$

Arguments

canonical_model::CanonicalModel : the canonical model built in PowerSimulations

ts_data::Tuple{Vector{String}, Vector{Float64}, Vector{Vector{Float64}}} : timeseries data name (1), multiplier (2) and values (3)

cons_name::Symbol : name of the constraint

var_name::Symbol : the name of the variable

PowerSimulations.device_timeseries_ub_bigM — Function.

device_timeseries_ub_bigM(canonical_model::CanonicalModel,
                                ts_data::Tuple{Vector{String}, Vector{Float64}, Vector{Vector{Float64}}},
                                cons_name::Symbol,
                                var_name::Symbol,
                                param_reference::UpdateRef,
                                binvar_name::Symbol,
                                M_value::Float64 = 1e6)

Constructs upper bound for variable and time series and a multiplier or confines to 0 depending on binary variable. Uses BigM constraint type to allow for parameter since ParameterJuMP doesn't support var*parameter

Constraints

varcts[name, t] - ts_data[2][ix] * param[name, t] <= (1 - varbin[name, t])*M_value

varcts[name, t] <= varbin[name, t]*M_value

LaTeX

$x^{cts}_t - r^{val} x^{param}_t \leq M(1 - x^{bin}_t ), forall t$

$x^{cts}_t \leq M x^{bin}_t, \forall t$

Arguments

canonical_model::CanonicalModel : the canonical model built in PowerSimulations

ts_data::Tuple{Vector{String}, Vector{Vector{Float64}}} : timeseries data name (1) and values (2)

cons_name::Symbol : name of the constraint

var_name::Symbol : name of the variable

param_reference::UpdateRef : UpdateRef of access the parameters

binvar_name::Symbol : name of binary variable

M_value::Float64 : bigM

PowerSimulations.device_timeseries_ub_bin — Method.

device_timeseries_ub_bin(canonical_model::CanonicalModel,
                                ts_data::Tuple{Vector{String}, Vector{Float64}, Vector{Vector{Float64}}},
                                cons_name::Symbol,
                                var_name::Symbol,
                                binvar_name::Symbol)

Constructs upper bound for variable and time series or confines to 0 depending on binary variable. The upper bound is defined by a time series and a multiplier.

Constraints

varcts[name, t] <= varbin[name, t]* ts_data[2][ix] * ts_data[3][ix][t]

where (ix, name) in enumerate(ts_data[1]).

LaTeX

$x^{cts}_t \leq r^{val} r_t x^{bin}_t, \forall t$

Arguments

canonical_model::CanonicalModel : the canonical model built in PowerSimulations

ts_data::Tuple{Vector{String}, Vector{Vector{Float64}}} : timeseries data name (1) and values (2)

cons_name::Symbol : name of the constraint

var_name::Symbol : name of the variable

binvar_name::Symbol : name of binary variable

PowerSimulations.energy_balance — Method.

energy_balance(canonical_model::CanonicalModel,
                    initial_conditions::Vector{InitialCondition},
                    efficiency_data::Tuple{Vector{String}, Vector{InOut}},
                    cons_name::Symbol,
                    var_names::Tuple{Symbol, Symbol, Symbol})

Constructs multi-timestep constraint from initial condition, efficiency data, and variable tuple

Constraints

If t = 1:

varenergy[name, 1] == initial_conditions[ix].value + varin[name, 1]*eff_in*fraction_of_hour - varout[name, 1]*fraction_of_hour/eff_out

If t > 1:

varenergy[name, t] == varenergy[name, t-1] + varin[name, t]*eff_in*fraction_of_hour - varout[name, t]*fraction_of_hour/eff_out

LaTeX

$x^{energy}_1 == x^{energy}_{init} + frhr \eta^{in} x^{in}_1 - \frac{frhr}{\eta^{out}} x^{out}_1, \text{ for } t = 1$

$x^{energy}_t == x^{energy}_{t-1} + frhr \eta^{in} x^{in}_t - \frac{frhr}{\eta^{out}} x^{out}_t, \forall t \geq 2$

Arguments

canonical_model::CanonicalModel : the canonical model built in PowerSimulations

initial_conditions::Vector{InitialCondition} : for time zero 'varenergy'

efficiency_data::Tuple{Vector{String}, Vector{InOut}} :: charging/discharging efficiencies

cons_name::Symbol : name of the constraint

var_names::Tuple{Symbol, Symbol, Symbol} : the names of the variables

: var_names[1] : varin

: var_names[2] : varout

: var_names[3] : varenergy

PowerSimulations.parameter_update! — Method.

Updates the forecast parameter value

PowerSimulations.post_nip — Method.

PowerSimulations.post_nip_expr — Method.

PowerSimulations.powermodels_network! — Method.

PowerSimulations.powermodels_network! — Method.

PowerSimulations.ps_cost — Method.

ps_cost(canonical_model::CanonicalModel,
            variable::JuMP.Containers.DenseAxisArray{JV},
            cost_component::Float64,
            dt::Float64,
            sign::Float64) where {JV <: JuMP.AbstractVariableRef}

Returns linear cost terms for sum of variables with common factor to be used for cost expression for canonical model.

Equation

gen_cost = sum(variable)*cost_component

LaTeX

$cost = dt\times sign\sum_{i\in I} c x_i$

Returns:

sign*gen_cost*dt

Arguments

canonical_model::CanonicalModel : the canonical model built in PowerSimulations

variable::JuMP.Containers.DenseAxisArray{JV} : variable array

cost_component::Float64 : cost to be associated with variable

dt::Float64 : fraction of hour

sign::Float64 : positive or negative sign to be associated cost term

PowerSimulations.ps_cost — Method.

ps_cost(canonical_model::CanonicalModel,
             variable::JuMP.Containers.DenseAxisArray{JV},
             cost_component::PSY.VariableCost{Vector{NTuple{2, Float64}}},
             dt::Float64,
             sign::Float64) where {JV<:JuMP.AbstractVariableRef}

Creates piecewise linear cost function using a sum of variables and expression with sign and time step included.

Expression

JuMP.add_to_expression!(gen_cost, c)

Returns signgen_costdt

LaTeX

$cost = sign\times dt \sum_{v\in V} c_v$

where $c_v$ is given by

$c_v = \sum_{i\in Ix} \frac{y_i - y_{i-1}}{x_i - x_{i-1}} v^{p.w.}_i$

Arguments

canonical_model::CanonicalModel : the canonical model built in PowerSimulations

variable::JuMP.Containers.DenseAxisArray{JV} : variable array

cost_component::PSY.VariableCost{Vector{NTuple{2, Float64}}}

dt::Float64 : fraction of hour

sign::Float64 : positive or negative sign to be associated cost term

PowerSimulations.ps_cost — Method.

ps_cost(canonical_model::CanonicalModel,
            variable::JuMP.Containers.DenseAxisArray{JV},
            cost_component::PSY.VariableCost{Float64},
            dt::Float64,
            sign::Float64) where {JV <: JuMP.AbstractVariableRef}

Returns linear cost terms for sum of variables with common factor to be used for cost expression for canonical model. Does this by calling ps_cost that has Float64 cost component input.

Returns:

ps_cost(canonical_model, variable, PSY.get_cost(cost_component), dt, sign)

Arguments

canonical_model::CanonicalModel : the canonical model built in PowerSimulations

variable::JuMP.Containers.DenseAxisArray{JV} : variable array

cost_component::PSY.VariableCost{Float64} : container for cost to be associated with variable

dt::Float64 : fraction of hour

sign::Float64 : positive or negative sign to be associated cost term

PowerSimulations.ps_cost — Method.

ps_cost(canonical_model::CanonicalModel,
            variable::JuMP.Containers.DenseAxisArray{JV},
            cost_component::PSY.VariableCost{NTuple{2, Float64}}
            dt::Float64,
            sign::Float64) where {JV <: JuMP.AbstractVariableRef}

Returns quadratic cost terms for sum of variables with common factor to be used for cost expression for canonical model.

Equation

gen_cost = dt*sign*(sum(variable.^2)*cost_component[1] + sum(variable)*cost_component[2])

LaTeX

$cost = dt\times sign (sum_{i\in I} c_1 v_i^2 + sum_{i\in I} c_2 v_i )$

for quadratic factor large enough. Otherwise

return ps_cost(canonical_model, variable, cost_component[2], dt, 1.0)

Returns gen_cost

Arguments

canonical_model::CanonicalModel : the canonical model built in PowerSimulations

variable::JuMP.Containers.DenseAxisArray{JV} : variable array

cost_component::PSY.VariableCost{NTuple{2, Float64}} : container for quadratic and linear factors

sign::Float64 : positive or negative sign to be associated cost term

PowerSimulations.ramp_constraints! — Method.

This function adds the ramping limits of generators when there are CommitmentVariables

PowerSimulations.range_ff — Method.

    range_ff(canonical_model::CanonicalModel,
                    cons_name::Symbol,
                    param_reference::NTuple{2, UpdateRef},
                    var_name::Symbol)

Constructs min/max range parametrized constraint from device variable to include feedforward.

Constraints

param_reference[1][var_name] <= variable[var_name, t] variable[var_name, t] <= param_reference[2][var_name]

where r in range_data.

LaTeX

$param^{min} \leq x$ $x \leq param^{max}$

Arguments

canonical_model::CanonicalModel : the canonical model built in PowerSimulations

param_reference::NTuple{2, UpdateRef} : Tuple with the lower bound and upper bound parameter reference

cons_name::Symbol : name of the constraint

var_name::Symbol : the name of the continuous variable

PowerSimulations.rating_constraint — Method.

rating_constraint(canonical_model::CanonicalModel,
                        rating_data::Vector{Tuple{String, Float64}},
                        cons_name::Symbol,
                        var_names::Tuple{Symbol, Symbol})

Constructs constraint from rating data and related variable tuple.

Constraint

var1[r[1], t]^2 + var2[r[1], t]^2 <= r[2]^2

where r in rating data and t in time steps.

LaTeX

$x_1^2 + x_2^2 \leq r^2$

Arguments

canonical_model::CanonicalModel : the canonical model built in PowerSimulations

rating_data::Vector{Tuple{String, Float64}} : rating data name (1) and value (2)

cons_name::Symbol : name of the constraint

var_names::Tuple{Symbol, Symbol} : the names of the variables

: var_names[1] : var1

: var_names[2] : var2

PowerSimulations.reactive_power_constraints — Method.

This function adds the reactive power limits of generators when there are CommitmentVariables

PowerSimulations.reactivepower_constraints! — Method.

This function adds the reactive power limits of generators when there are CommitmentVariables

PowerSimulations.reactivepower_constraints! — Method.

This function adds the reactive power limits of generators when there CommitmentVariables

PowerSimulations.reactivepower_constraints — Method.

Reactive Power Constraints on Loads Assume Constant PowerFactor

PowerSimulations.reactivepower_variables! — Method.

This function add the variables for power generation output to the model

PowerSimulations.remove_chars — Method.

Removes the string char from the original string

PowerSimulations.replace_chars — Method.

Replaces the string in char with the stringreplacement

PowerSimulations.reserve_device_semicontinuousrange — Method.

reserve_device_semicontinuousrange(canonical_model::CanonicalModel,
                                scrange_data::Vector{NamedMinMax},
                                cons_name::Symbol,
                                var_name::Symbol,
                                binvar_name::Symbol)

Constructs min/max range constraint from device variable and on/off decision variable.

Constraints

If device min = 0:

varcts[r[1], t] <= r[2].max*(1-varbin[r[1], t])

varcts[r[1], t] >= 0.0

Otherwise:

varcts[r[1], t] <= r[2].max*(1-varbin[r[1], t])

varcts[r[1], t] >= r[2].min*(1-varbin[r[1], t])

where r in range_data.

LaTeX

$0 \leq x^{cts} \leq r^{max} (1 - x^{bin} ), \text{ for } r^{min} = 0$

$r^{min} (1 - x^{bin} ) \leq x^{cts} \leq r^{max} (1 - x^{bin} ), \text{ otherwise }$

Arguments

canonical_model::CanonicalModel : the canonical model built in PowerSimulations

scrange_data::Vector{NamedMinMax} : contains name of device (1) and its min/max (2)

cons_name::Symbol : name of the constraint

var_name::Symbol : the name of the continuous variable

binvar_name::Symbol : the name of the binary variable

PowerSimulations.semicontinuousrange_ff — Method.

        semicontinuousrange_ff(canonical_model::CanonicalModel,
                                cons_name::Symbol,
                                var_name::Symbol,
                                param_reference::UpdateRef)

Constructs min/max range constraint from device variable with parameter setting.

Constraints

If device min = 0:

variable[var_name, t] <= r[2].max*param_reference[var_name]

Otherwise:

variable[var_name, t] <= r[2].max*param_reference[var_name]

variable[var_name, t] >= r[2].min*param_reference[var_name]

where r in range_data.

LaTeX

$0.0 \leq x^{var} \leq r^{max} x^{param}, \text{ for } r^{min} = 0$

$r^{min} x^{param} \leq x^{var} \leq r^{min} x^{param}, \text{ otherwise }$

Arguments

canonical_model::CanonicalModel : the canonical model built in PowerSimulations

cons_name::Symbol : name of the constraint

var_name::Symbol : the name of the continuous variable

param_reference::UpdateRef : UpdateRef of the parameter

PowerSimulations.set_variable_bounds — Method.

set_variable_bounds(canonical_model::CanonicalModel,
                        bounds::Vector{NamedMinMax},
                        var_name::Symbol)

Adds a bounds to a variable in the optimization model.

Bounds

bounds.min <= varstart[name, t] <= bounds.max

LaTeX

$x^{device}_t >= bound^{min} \forall t$

$x^{device}_t <= bound^{max} \forall t$

Arguments

canonical_model::CanonicalModel : the canonical model built in PowerSimulations

bounds::Vector{NamedMinMax} : contains name of device (1) and its min/max (2)

var_name::Symbol : Base Name for the variable

PowerSimulations.status_init — Method.

Status Initis is always calculated based on the Power Output of the device This is to make it easier to calculate when the previous model doesn't contain binaries. For instance, looking back on an ED model to find the IC of the UC model

PowerSimulations.ub_ff — Method.

    ub_ff(canonical_model::CanonicalModel,
          cons_name::Symbol,
          param_reference::UpdateRef,
          var_name::Symbol)

Constructs a parametrized upper bound constraint to implement feedforward from other models. The Parameters are initialized using the uppper boundary values of the provided variables.

Constraints

variable[var_name, t] <= param_reference[var_name]

LaTeX

$x \leq param^{max}$

Arguments

canonical_model::CanonicalModel : the canonical model built in PowerSimulations

cons_name::Symbol : name of the constraint

param_reference : Reference to the Parameter used to determine the upperbound

var_name::Symbol : the name of the continuous variable

PowerSimulations.variable_active_net_injection — Method.

PowerSimulations.variable_net_injection — Method.

generates variables for both active and reactive net injection

PowerSimulations.variable_reactive_net_injection — Method.

PowerSimulations.variable_reactive_net_injection — Method.

active power only models ignore reactive power variables