PowerSystems.Branch Formulations

Valid DeviceModels for subtypes of Branch include the following:

Valid DeviceModel	Device Type	Formulation

DeviceModel(HVDCLine, HVDCDispatch)	HVDCLine	HVDCDispatch

DeviceModel(VSCDCLine, HVDCDispatch)	VSCDCLine	HVDCDispatch

DeviceModel(HVDCLine, HVDCLossless)	HVDCLine	HVDCLossless

DeviceModel(VSCDCLine, HVDCLossless)	VSCDCLine	HVDCLossless

DeviceModel(HVDCLine, HVDCUnbounded)	HVDCLine	HVDCUnbounded

DeviceModel(VSCDCLine, HVDCUnbounded)	VSCDCLine	HVDCUnbounded

DeviceModel(DynamicBranch, StaticBranch)	DynamicBranch	StaticBranch

DeviceModel(Line, StaticBranch)	Line	StaticBranch

DeviceModel(MonitoredLine, StaticBranch)	MonitoredLine	StaticBranch

DeviceModel(PhaseShiftingTransformer, StaticBranch)	PhaseShiftingTransformer	StaticBranch

DeviceModel(TapTransformer, StaticBranch)	TapTransformer	StaticBranch

DeviceModel(Transformer2W, StaticBranch)	Transformer2W	StaticBranch

DeviceModel(DynamicBranch, StaticBranchBounds)	DynamicBranch	StaticBranchBounds

DeviceModel(Line, StaticBranchBounds)	Line	StaticBranchBounds

DeviceModel(MonitoredLine, StaticBranchBounds)	MonitoredLine	StaticBranchBounds

DeviceModel(PhaseShiftingTransformer, StaticBranchBounds)	PhaseShiftingTransformer	StaticBranchBounds

DeviceModel(TapTransformer, StaticBranchBounds)	TapTransformer	StaticBranchBounds

DeviceModel(Transformer2W, StaticBranchBounds)	Transformer2W	StaticBranchBounds

DeviceModel(DynamicBranch, StaticBranchUnbounded)	DynamicBranch	StaticBranchUnbounded

DeviceModel(Line, StaticBranchUnbounded)	Line	StaticBranchUnbounded

DeviceModel(MonitoredLine, StaticBranchUnbounded)	MonitoredLine	StaticBranchUnbounded

DeviceModel(PhaseShiftingTransformer, StaticBranchUnbounded)	PhaseShiftingTransformer	StaticBranchUnbounded

DeviceModel(TapTransformer, StaticBranchUnbounded)	TapTransformer	StaticBranchUnbounded

DeviceModel(Transformer2W, StaticBranchUnbounded)	Transformer2W	StaticBranchUnbounded

StaticBranch

PowerSimulations.StaticBranch — Type

Branch type to add unbounded flow variables and use flow constraints

StaticBranch()

defined at /home/runner/work/PowerSimulations.jl/PowerSimulations.jl/src/core/formulations.jl:176.

StaticBranchBounds

PowerSimulations.StaticBranchBounds — Type

Branch type to add bounded flow variables and use flow constraints

StaticBranchBounds()

defined at /home/runner/work/PowerSimulations.jl/PowerSimulations.jl/src/core/formulations.jl:180.

StaticBranchUnbounded

PowerSimulations.StaticBranchUnbounded — Type

Branch type to avoid flow constraints

StaticBranchUnbounded()

defined at /home/runner/work/PowerSimulations.jl/PowerSimulations.jl/src/core/formulations.jl:184.

HVDCLossless

PowerSimulations.HVDCLossless — Type

Branch type to represent lossless power flow on DC lines

HVDCLossless()

defined at /home/runner/work/PowerSimulations.jl/PowerSimulations.jl/src/core/formulations.jl:195.

HVDCDispatch

PowerSimulations.HVDCDispatch — Type

Branch type to represent lossy power flow on DC lines

HVDCDispatch()

defined at /home/runner/work/PowerSimulations.jl/PowerSimulations.jl/src/core/formulations.jl:199.

HVDCUnbounded

PowerSimulations.HVDCUnbounded — Type

Branch type to avoid flow constraints

HVDCUnbounded()

defined at /home/runner/work/PowerSimulations.jl/PowerSimulations.jl/src/core/formulations.jl:191.