PowerSystems.Branch Formulations

Note

The use of reactive power variables and constraints will depend on the network model used, i.e., whether it uses (or does not use) reactive power. If the network model is purely active power-based, reactive power variables and related constraints are not created.

Table of contents

StaticBranch

StaticBranchBounds

StaticBranchUnbounded

HVDCTwoTerminalUnbounded

HVDCTwoTerminalLossless

HVDCTwoTerminalDispatch

PhaseAngleControl

TwoTerminalLCCLine

Valid configurations

StaticBranch

Formulation valid for PTDFPowerModel Network model

PowerSimulations.StaticBranch — Type

Branch type to add unbounded flow variables and use flow constraints

Variables:

FlowActivePowerVariable:

Bounds: $(-\infty,\infty)$

Symbol: $f$ If Slack variables are enabled:

FlowActivePowerSlackUpperBound:

Bounds: [0.0, ]

Default proportional cost: 2e5

Symbol: $f^\text{sl,up}$

FlowActivePowerSlackLowerBound:

Bounds: [0.0, ]

Default proportional cost: 2e5

Symbol: $f^\text{sl,lo}$

Static Parameters

$R^\text{max}$ = PowerSystems.get_rating(branch)

Objective:

Add a large proportional cost to the objective function if rate constraint slack variables are used $+ (f^\text{sl,up} + f^\text{sl,lo}) \cdot 2 \cdot 10^5$

Expressions:

No expressions are used.

Constraints:

For each branch $b \in \{1,\dots, B\}$ (in a system with $N$ buses) the constraints are given by:

\[\begin{aligned} & f_t = \sum_{i=1}^N \text{PTDF}_{i,b} \cdot \text{Bal}_{i,t}, \quad \forall t \in \{1,\dots, T\}\\ & f_t - f_t^\text{sl,up} \le R^\text{max},\quad \forall t \in \{1,\dots, T\} \\ & f_t + f_t^\text{sl,lo} \ge -R^\text{max},\quad \forall t \in \{1,\dots, T\} \end{aligned}\]

on which $\text{PTDF}$ is the $N \times B$ system Power Transfer Distribution Factors (PTDF) matrix, and $\text{Bal}_{i,t}$ is the active power bus balance expression (i.e. $\text{Generation}_{i,t} - \text{Demand}_{i,t}$) at bus $i$ at time-step $t$.

StaticBranchBounds

Formulation valid for PTDFPowerModel Network model

PowerSimulations.StaticBranchBounds — Type

Branch type to add bounded flow variables and use flow constraints

Variables:

FlowActivePowerVariable:

Bounds: $\left[-R^\text{max},R^\text{max}\right]$

Symbol: $f$

Static Parameters

$R^\text{max}$ = PowerSystems.get_rating(branch)

Objective:

No cost is added to the objective function.

Expressions:

No expressions are used.

Constraints:

For each branch $b \in \{1,\dots, B\}$ (in a system with $N$ buses) the constraints are given by:

\[\begin{aligned} & f_t = \sum_{i=1}^N \text{PTDF}_{i,b} \cdot \text{Bal}_{i,t}, \quad \forall t \in \{1,\dots, T\} \end{aligned}\]

on which $\text{PTDF}$ is the $N \times B$ system Power Transfer Distribution Factors (PTDF) matrix, and $\text{Bal}_{i,t}$ is the active power bus balance expression (i.e. $\text{Generation}_{i,t} - \text{Demand}_{i,t}$) at bus $i$ at time-step $t$.

StaticBranchUnbounded

Formulation valid for PTDFPowerModel Network model

PowerSimulations.StaticBranchUnbounded — Type

Branch type to avoid flow constraints

FlowActivePowerVariable:

Bounds: $(-\infty,\infty)$

Symbol: $f$

Objective:

No cost is added to the objective function.

Expressions:

No expressions are used.

Constraints:

For each branch $b \in \{1,\dots, B\}$ (in a system with $N$ buses) the constraints are given by:

\[\begin{aligned} & f_t = \sum_{i=1}^N \text{PTDF}_{i,b} \cdot \text{Bal}_{i,t}, \quad \forall t \in \{1,\dots, T\} \end{aligned}\]

on which $\text{PTDF}$ is the $N \times B$ system Power Transfer Distribution Factors (PTDF) matrix, and $\text{Bal}_{i,t}$ is the active power bus balance expression (i.e. $\text{Generation}_{i,t} - \text{Demand}_{i,t}$) at bus $i$ at time-step $t$.

HVDCTwoTerminalUnbounded

Formulation valid for PTDFPowerModel Network model

PowerSimulations.HVDCTwoTerminalUnbounded — Type

Branch type to avoid flow constraints

This model assumes that it can transfer power from two AC buses without losses and no limits.

Variables:

FlowActivePowerVariable:

Bounds: $\left(-\infty,\infty\right)$

Symbol: $f$

Objective:

No cost is added to the objective function.

Expressions:

The variable FlowActivePowerVariable $f$ is added to the nodal balance expression ActivePowerBalance, by adding the flow $f$ in the receiving bus and subtracting it from the sending bus. This is used then to compute the AC flows using the PTDF equation.

Constraints:

No constraints are added.

HVDCTwoTerminalLossless

Formulation valid for PTDFPowerModel Network model

PowerSimulations.HVDCTwoTerminalLossless — Type

Branch type to represent lossless power flow on DC lines

This model assumes that it can transfer power from two AC buses without losses.

Variables:

FlowActivePowerVariable:

Bounds: $\left(-\infty,\infty\right)$

Symbol: $f$

Static Parameters

$R^\text{from,min}$ = PowerSystems.get_active_power_limits_from(branch).min

$R^\text{from,max}$ = PowerSystems.get_active_power_limits_from(branch).max

$R^\text{to,min}$ = PowerSystems.get_active_power_limits_to(branch).min

$R^\text{to,max}$ = PowerSystems.get_active_power_limits_to(branch).max

Objective:

No cost is added to the objective function.

Expressions:

The variable FlowActivePowerVariable $f$ is added to the nodal balance expression ActivePowerBalance, by adding the flow $f$ in the receiving bus and subtracting it from the sending bus. This is used then to compute the AC flows using the PTDF equation.

Constraints:

\[\begin{align*} & R^\text{min} \le f_t \le R^\text{max},\quad \forall t \in \{1,\dots, T\} \\ \end{align*}\]

where:

\[\begin{align*} & R^\text{min} = \begin{cases} \min\left(R^\text{from,min}, R^\text{to,min}\right), & \text{if } R^\text{from,min} \ge 0 \text{ and } R^\text{to,min} \ge 0 \\ \max\left(R^\text{from,min}, R^\text{to,min}\right), & \text{if } R^\text{from,min} \le 0 \text{ and } R^\text{to,min} \le 0 \\ R^\text{from,min},& \text{if } R^\text{from,min} \le 0 \text{ and } R^\text{to,min} \ge 0 \\ R^\text{to,min},& \text{if } R^\text{from,min} \ge 0 \text{ and } R^\text{to,min} \le 0 \end{cases} \end{align*}\]

and

\[\begin{align*} & R^\text{max} = \begin{cases} \min\left(R^\text{from,max}, R^\text{to,max}\right), & \text{if } R^\text{from,max} \ge 0 \text{ and } R^\text{to,max} \ge 0 \\ \max\left(R^\text{from,max}, R^\text{to,max}\right), & \text{if } R^\text{from,max} \le 0 \text{ and } R^\text{to,max} \le 0 \\ R^\text{from,max},& \text{if } R^\text{from,max} \le 0 \text{ and } R^\text{to,max} \ge 0 \\ R^\text{to,max},& \text{if } R^\text{from,max} \ge 0 \text{ and } R^\text{to,max} \le 0 \end{cases} \end{align*}\]

HVDCTwoTerminalDispatch

Formulation valid for PTDFPowerModel Network model

PowerSimulations.HVDCTwoTerminalDispatch — Type

Branch type to represent lossy power flow on DC lines

Variables

FlowActivePowerToFromVariable:

Symbol: $f^\text{to-from}$

FlowActivePowerFromToVariable:

Symbol: $f^\text{from-to}$

HVDCLosses:

Symbol: $\ell$

HVDCFlowDirectionVariable

Bounds: $\{0,1\}$

Symbol: $u^\text{dir}$

Static Parameters

$R^\text{from,min}$ = PowerSystems.get_active_power_limits_from(branch).min

$R^\text{from,max}$ = PowerSystems.get_active_power_limits_from(branch).max

$R^\text{to,min}$ = PowerSystems.get_active_power_limits_to(branch).min

$R^\text{to,max}$ = PowerSystems.get_active_power_limits_to(branch).max

$L_0$ = PowerSystems.get_loss(branch).l0

$L_1$ = PowerSystems.get_loss(branch).l1

Objective:

No cost is added to the objective function.

Expressions:

Each FlowActivePowerToFromVariable $f^\text{to-from}$ and FlowActivePowerFromToVariable $f^\text{from-to}$ is added to the nodal balance expression ActivePowerBalance, by adding the respective flow in the receiving bus and subtracting it from the sending bus. That is, $f^\text{to-from}$ adds the flow to the from bus, and subtracts the flow from the to bus, while $f^\text{from-to}$ adds the flow to the to bus, and subtracts the flow from the from bus This is used then to compute the AC flows using the PTDF equation.

In addition, the HVDCLosses are subtracted to the from bus in the ActivePowerBalance expression.

Constraints:

\[\begin{align*} & R^\text{from,min} \le f_t^\text{from-to} \le R^\text{from,max}, \forall t \in \{1,\dots, T\} \\ & R^\text{to,min} \le f_t^\text{to-from} \le R^\text{to,max},\quad \forall t \in \{1,\dots, T\} \\ & f_t^\text{to-from} - f_t^\text{from-to} \le L_1 \cdot f_t^\text{to-from} - L_0,\quad \forall t \in \{1,\dots, T\} \\ & f_t^\text{from-to} - f_t^\text{to-from} \ge L_1 \cdot f_t^\text{from-to} + L_0,\quad \forall t \in \{1,\dots, T\} \\ & f_t^\text{from-to} - f_t^\text{to-from} \ge - M^\text{big} (1 - u^\text{dir}_t),\quad \forall t \in \{1,\dots, T\} \\ & f_t^\text{to-from} - f_t^\text{from-to} \ge - M^\text{big} u^\text{dir}_t,\quad \forall t \in \{1,\dots, T\} \\ & f_t^\text{to-from} - f_t^\text{from-to} \le \ell_t,\quad \forall t \in \{1,\dots, T\} \\ & f_t^\text{from-to} - f_t^\text{to-from} \le \ell_t,\quad \forall t \in \{1,\dots, T\} \end{align*}\]

PhaseAngleControl

Formulation valid for PTDFPowerModel Network model

PowerSimulations.PhaseAngleControl — Type

Branch formulation for PhaseShiftingTransformer flow control

Variables:

FlowActivePowerVariable:

Bounds: $(-\infty,\infty)$

Symbol: $f$

PhaseShifterAngle:

Symbol: $\theta^\text{shift}$

Static Parameters

$R^\text{max}$ = PowerSystems.get_rating(branch)

$\Theta^\text{min}$ = PowerSystems.get_phase_angle_limits(branch).min

$\Theta^\text{max}$ = PowerSystems.get_phase_angle_limits(branch).max

$X$ = PowerSystems.get_x(branch) (series reactance)

Objective:

No changes to objective function

Expressions:

Adds to the ActivePowerBalance expression the term $-\theta^\text{shift} /X$ to the from bus and $+\theta^\text{shift} /X$ to the to bus, that the PhaseShiftingTransformer is connected.

Constraints:

For each branch $b \in \{1,\dots, B\}$ (in a system with $N$ buses) the constraints are given by:

\[\begin{aligned} & f_t = \sum_{i=1}^N \text{PTDF}_{i,b} \cdot \text{Bal}_{i,t} + \frac{\theta^\text{shift}_t}{X}, \quad \forall t \in \{1,\dots, T\}\\ & -R^\text{max} \le f_t \le R^\text{max},\quad \forall t \in \{1,\dots, T\} \end{aligned}\]

on which $\text{PTDF}$ is the $N \times B$ system Power Transfer Distribution Factors (PTDF) matrix, and $\text{Bal}_{i,t}$ is the active power bus balance expression (i.e. $\text{Generation}_{i,t} - \text{Demand}_{i,t}$) at bus $i$ at time-step $t$.

TwoTerminalLCCLine

Formulation valid for ACPPowerModel Network model

Variables:

[HVDCRectifierDelayAngleVariable]:

+ Bounds: ``(-\alpha_{r,t}^\text{min},\alpha_{r,t}^\text{max})``
+ Symbol: ``\alpha_{r,t}``

[HVDCInverterExtinctionAngleVariable]:

+ Bounds: ``(-\gamma_{i,t}^\text{min},\gamma_{i,t}^\text{max})``
+ Symbol: ``\gamma_{i,t}``

[HVDCRectifierPowerFactorAngleVariable]:

+ Bounds: ``\{0,1\}``
+ Symbol: ``\phi_{r,t}``

[HVDCInverterPowerFactorAngleVariable]:

+ Bounds: ``\{0,1\}``
+ Symbol: ``\phi_{i,t}``

[HVDCRectifierOverlapAngleVariable]:

+ Bounds: [0.0, ]
+ Symbol: ``\mu_{r,t}``

[HVDCInverterOverlapAngleVariable]:

+ Bounds: [0.0, ]
+ Symbol: ``\mu_{i,t}``

[HVDCRectifierTapSettingVariable]:

+ Bounds: ``(t_{r,t}^\text{min},t_{r,t}^\text{max})``
+ Symbol: ``t_{r,t}``

[HVDCInverterTapSettingVariable]:

+ Bounds: ``(t_{i,t}^\text{min},t_{i,t}^\text{max})``
+ Symbol: ``t_{i,t}``

[HVDCRectifierDCVoltageVariable]:

+ Bounds: [0.0, ]
+ Symbol: ``v_{r,t}^\text{dc}``

[HVDCInverterDCVoltageVariable]:

+ Bounds: [0.0, ]
+ Symbol: ``v_{i,t}^\text{dc}``

[HVDCRectifierACCurrentVariable]:

+ Bounds: [0.0, ]
+ Symbol: ``I_{r,t}^\text{ac}``

[HVDCInverterACCurrentVariable]:

+ Bounds: [0.0, ]
+ Symbol: ``I_{i,t}^\text{ac}``

[DCLineCurrentFlowVariable]:

+ Bounds: [0.0, ]
+ Symbol: ``I^\text{dc}``

[HVDCActivePowerReceivedFromVariable]:

+ Bounds: [0.0, ]
+ Symbol: ``p_{r,t}^\text{ac}``

[HVDCActivePowerReceivedToVariable]:

+ Bounds: [0.0, ]
+ Symbol: ``p_{i,t}^\text{ac}``

[HVDCReactivePowerReceivedFromVariable]:

+ Bounds: [0.0, ]
+ Symbol: ``q_{r,t}^\text{ac}``

[HVDCReactivePowerReceivedToVariable]:

+ Bounds: [0.0, ]
+ Symbol: ``q_{i,t}^\text{ac}``

Static Parameters

$R^\text{dc}$ = PowerSystems.get_r(lcc)

$N_r$ = PowerSystems.get_rectifier_bridges(lcc)

$N_i$ = PowerSystems.get_inverter_bridges(lcc)

$X_r$ = PowerSystems.get_rectifier_xc(lcc)

$X_i$ = PowerSystems.get_inverter_xc(lcc)

$a_r$ = PowerSystems.get_rectifier_transformer_ratio(lcc)

$a_i$ = PowerSystems.get_inverter_transformer_ratio(lcc)

$t_r$ = PowerSystems.get_rectifier_tap_setting(lcc)

$t_i$ = PowerSystems.get_inverter_tap_setting(lcc)

$t^\text{min}_r$ = PowerSystems.get_rectifier_tap_setting(lcc).min

$t^\text{max}_r$ = PowerSystems.get_rectifier_tap_setting(lcc).max

$t^\text{min}_i$ = PowerSystems.get_inverter_tap_setting(lcc).min

$t^\text{max}_i$ = PowerSystems.get_inverter_tap_setting(lcc).max

Objective:

No changes to objective function

Expressions:

The variable HVDCActivePowerReceivedFromVariable $p_{r,t}^\text{ac}$ is added to the nodal balance expression ActivePowerBalance as a negative load, since the rectifier takes power from the AC system and to injects it into the DC system. On the other hand, the variable HVDCActivePowerReceivedToVariable $p_{i,t}^\text{ac}$ is added to the nodal balance expression ActivePowerBalance as a positive load, since it takes the power from the DC system and injects it back into the AC system.

The variables HVDCReactivePowerReceivedFromVariable $q_{r,t}^\text{ac}$ and HVDCReactivePowerReceivedToVariable `q_{i,t}^\text{ac}are added to the nodal balance expressionActivePowerBalance as positive loads, since they consume reactive power from the AC system to allow current transfer in converters during commutation.

Constraints:

Rectifier:

\[\begin{aligned} & v^\text{dc}_{r,t} = \frac{3}{\pi}N_r \left( \sqrt{2}\frac{a_r v^\text{ac}_{r,t}}{t_{r,t}}\\cos{\alpha_{r,t}}-X_r I^\text{dc}_t \right)\\ & \mu_{r,t} = \arccos \left( \cos\alpha_{r,t} - \frac{\sqrt{2} I^\text{dc}_t X_r t_{r,t}}{a_r v^\text{ac}_{r,t}} \right) - \alpha_{r,t}\\ & \phi_{r,t} = \arctan \left( \frac{2\mu_{r,t} + \sin(2\alpha_{r,t}) - \sin(2\mu_{r,t} + 2\alpha_{r,t})}{\cos(2\alpha_{r,t}) - \cos(2\mu_{r,t} + 2\alpha_{r,t})} \right)\\ \end{aligned}\]

Which can be approximated as:

\[\begin{aligned} & \phi_{r,t} = arccos(\frac{1}{2}\cos\alpha_{r,t} + \frac{1}{2}\cos(\alpha_{r,t} + \mu_{r,t})) \end{aligned}\]

\[\begin{aligned} & I^\text{ac}_{r,t} = \sqrt{6} \frac{N_r}{\pi} I^\text{dc}_t\\ & p^\text{ac}_{r,t} = \sqrt{3} I^\text{ac}_{r,t} \frac{a_r v^\text{ac}_{r,t}}{t_{r,t}}\cos{\phi_{r,t}} \\ & q^\text{ac}_{r,t} = \sqrt{3} I^\text{ac}_{r,t} \frac{a_r v^\text{ac}_{r,t}}{t_{r,t}}\sin{\phi_{r,t}} \\ \end{aligned}\]

Inverter:

\[\begin{aligned} & v^\text{dc}_{i,t} = \frac{3}{\pi}N_i \left( \sqrt{2}\frac{a_i v^\text{ac}_{i,t}}{t_{i,t}}\\cos{\gamma_{i,t}}-X_i I^\text{dc}_t \right)\\ & \mu_{i,t} = \arccos \left( \cos\gamma_{i,t} - \frac{\sqrt{2} I^\text{dc}_t X_i t_{i,t}}{a_i v^\text{ac}_{i,t}} \right) - \gamma_{i,t}\\ & \phi_{i,t} = \arctan \left( \frac{2\mu_{i,t} + \sin(2\gamma_{i,t}) - \sin(2\mu_{i,t} + 2\gamma_{i,t})}{\cos(2\gamma_{i,t}) - \cos(2\mu_{r,t} + 2\gamma_{i,t})} \right)\\ \end{aligned}\]

Which can be approximated as:

\[\begin{aligned} & \phi_{i,t} = arccos(\frac{1}{2}\cos\gamma_{i,t} + \frac{1}{2}\cos(\gamma_{i,t} + \mu_{i,t})) \end{aligned}\]

\[\begin{aligned} & I^\text{ac}_{i,t} = \sqrt{6} \frac{N_i}{\pi} I^\text{dc}_t\\ & p^\text{ac}_{i,t} = \sqrt{3} I^\text{ac}_{i,t} \frac{a_i v^\text{ac}_{i,t}}{t_{i,t}}\cos{\phi_{i,t}} \\ & q^\text{ac}_{i,t} = \sqrt{3} I^\text{ac}_{i,t} \frac{a_i v^\text{ac}_{i,t}}{t_{i,t}}\sin{\phi_{i,t}} \\ \end{aligned}\]

DC Transmission Line:

\[\begin{aligned} & v^\text{dc}_{i,t} = v^\text{dc}_{r,t} - R_\text{dc}I^\text{dc}_t \end{aligned}\]

Valid configurations

Valid DeviceModels for subtypes of Branch include the following:

Valid DeviceModel	Device Type	Formulation

DeviceModel(PhaseShiftingTransformer, PhaseAngleControl)	PhaseShiftingTransformer	PhaseAngleControl

DeviceModel(TwoTerminalGenericHVDCLine, HVDCTwoTerminalDispatch)	TwoTerminalGenericHVDCLine	HVDCTwoTerminalDispatch

DeviceModel(TwoTerminalLCCLine, HVDCTwoTerminalDispatch)	TwoTerminalLCCLine	HVDCTwoTerminalDispatch

DeviceModel(TwoTerminalVSCLine, HVDCTwoTerminalDispatch)	TwoTerminalVSCLine	HVDCTwoTerminalDispatch

DeviceModel(TwoTerminalGenericHVDCLine, HVDCTwoTerminalLossless)	TwoTerminalGenericHVDCLine	HVDCTwoTerminalLossless

DeviceModel(TwoTerminalLCCLine, HVDCTwoTerminalLossless)	TwoTerminalLCCLine	HVDCTwoTerminalLossless

DeviceModel(TwoTerminalVSCLine, HVDCTwoTerminalLossless)	TwoTerminalVSCLine	HVDCTwoTerminalLossless

DeviceModel(TwoTerminalGenericHVDCLine, HVDCTwoTerminalUnbounded)	TwoTerminalGenericHVDCLine	HVDCTwoTerminalUnbounded

DeviceModel(TwoTerminalLCCLine, HVDCTwoTerminalUnbounded)	TwoTerminalLCCLine	HVDCTwoTerminalUnbounded

DeviceModel(TwoTerminalVSCLine, HVDCTwoTerminalUnbounded)	TwoTerminalVSCLine	HVDCTwoTerminalUnbounded

DeviceModel(TwoTerminalLCCLine, PowerSimulations.HVDCTwoTerminalLCC)	TwoTerminalLCCLine	PowerSimulations.HVDCTwoTerminalLCC

DeviceModel(TwoTerminalGenericHVDCLine, PowerSimulations.HVDCTwoTerminalPiecewiseLoss)	TwoTerminalGenericHVDCLine	PowerSimulations.HVDCTwoTerminalPiecewiseLoss

DeviceModel(TwoTerminalLCCLine, PowerSimulations.HVDCTwoTerminalPiecewiseLoss)	TwoTerminalLCCLine	PowerSimulations.HVDCTwoTerminalPiecewiseLoss

DeviceModel(TwoTerminalVSCLine, PowerSimulations.HVDCTwoTerminalPiecewiseLoss)	TwoTerminalVSCLine	PowerSimulations.HVDCTwoTerminalPiecewiseLoss

DeviceModel(DiscreteControlledACBranch, StaticBranch)	DiscreteControlledACBranch	StaticBranch

DeviceModel(Line, StaticBranch)	Line	StaticBranch

DeviceModel(MonitoredLine, StaticBranch)	MonitoredLine	StaticBranch

DeviceModel(PowerNetworkMatrices.BranchesSeries, StaticBranch)	PowerNetworkMatrices.BranchesSeries	StaticBranch

DeviceModel(PhaseShiftingTransformer3W, StaticBranch)	PhaseShiftingTransformer3W	StaticBranch

DeviceModel(Transformer3W, StaticBranch)	Transformer3W	StaticBranch

DeviceModel(PhaseShiftingTransformer, StaticBranch)	PhaseShiftingTransformer	StaticBranch

DeviceModel(TapTransformer, StaticBranch)	TapTransformer	StaticBranch

DeviceModel(Transformer2W, StaticBranch)	Transformer2W	StaticBranch

DeviceModel(AreaInterchange, StaticBranch)	AreaInterchange	StaticBranch

DeviceModel(DiscreteControlledACBranch, StaticBranchBounds)	DiscreteControlledACBranch	StaticBranchBounds

DeviceModel(Line, StaticBranchBounds)	Line	StaticBranchBounds

DeviceModel(MonitoredLine, StaticBranchBounds)	MonitoredLine	StaticBranchBounds

DeviceModel(PowerNetworkMatrices.BranchesSeries, StaticBranchBounds)	PowerNetworkMatrices.BranchesSeries	StaticBranchBounds

DeviceModel(PhaseShiftingTransformer3W, StaticBranchBounds)	PhaseShiftingTransformer3W	StaticBranchBounds

DeviceModel(Transformer3W, StaticBranchBounds)	Transformer3W	StaticBranchBounds

DeviceModel(PhaseShiftingTransformer, StaticBranchBounds)	PhaseShiftingTransformer	StaticBranchBounds

DeviceModel(TapTransformer, StaticBranchBounds)	TapTransformer	StaticBranchBounds

DeviceModel(Transformer2W, StaticBranchBounds)	Transformer2W	StaticBranchBounds

DeviceModel(DiscreteControlledACBranch, StaticBranchUnbounded)	DiscreteControlledACBranch	StaticBranchUnbounded

DeviceModel(Line, StaticBranchUnbounded)	Line	StaticBranchUnbounded

DeviceModel(MonitoredLine, StaticBranchUnbounded)	MonitoredLine	StaticBranchUnbounded

DeviceModel(PowerNetworkMatrices.BranchesSeries, StaticBranchUnbounded)	PowerNetworkMatrices.BranchesSeries	StaticBranchUnbounded

DeviceModel(PhaseShiftingTransformer3W, StaticBranchUnbounded)	PhaseShiftingTransformer3W	StaticBranchUnbounded

DeviceModel(Transformer3W, StaticBranchUnbounded)	Transformer3W	StaticBranchUnbounded

DeviceModel(PhaseShiftingTransformer, StaticBranchUnbounded)	PhaseShiftingTransformer	StaticBranchUnbounded

DeviceModel(TapTransformer, StaticBranchUnbounded)	TapTransformer	StaticBranchUnbounded

DeviceModel(Transformer2W, StaticBranchUnbounded)	Transformer2W	StaticBranchUnbounded

DeviceModel(AreaInterchange, StaticBranchUnbounded)	AreaInterchange	StaticBranchUnbounded