PowerSystems.ElectricLoad Formulations

Electric load formulations define the optimization models that describe load units (demand) mathematical model in different operational settings, such as economic dispatch and unit commitment.

Note

The use of reactive power variables and constraints will depend on the network model used, i.e., whether it uses (or does not use) reactive power. If the network model is purely active power-based, reactive power variables and related constraints are not created.

Table of contents

StaticPowerLoad

PowerLoadInterruption

PowerLoadDispatch

Valid configurations

StaticPowerLoad

PowerSimulations.StaticPowerLoad — Type

Formulation type to add a time series parameter for non-dispatchable ElectricLoad withdrawals to power balance constraints

Variables:

No variables are created

Time Series Parameters:

Uses the max_active_power timeseries parameter to determine the demand value at each time-step

Parameter	Default Time Series Name

ReactivePowerTimeSeriesParameter	max_active_power

ActivePowerTimeSeriesParameter	max_active_power

Expressions:

Subtracts the parameters listed above from the respective active and reactive power balance expressions created by the selected Network Formulations.

Constraints:

No constraints are created

PowerLoadInterruption

PowerSimulations.PowerLoadInterruption — Type

Formulation type to enable (binary) load interruptions

Variables:

ActivePowerVariable:

Bounds: [0.0, ]

Default initial value: 0.0

Symbol: $p^\text{ld}$

ReactivePowerVariable:

Bounds: [0.0, ]

Default initial value: 0.0

Symbol: $q^\text{ld}$

OnVariable:

Bounds: $\{0,1\}$

Default initial value: 1

Symbol: $u^\text{ld}$

Static Parameters:

$P^\text{ld,max}$ = PowerSystems.get_max_active_power(device)

$Q^\text{ld,max}$ = PowerSystems.get_max_reactive_power(device)

Time Series Parameters:

Parameter	Default Time Series Name

ReactivePowerTimeSeriesParameter	max_active_power

ActivePowerTimeSeriesParameter	max_active_power

Objective:

Creates an objective function term based on the FunctionData Options where the quantity term is defined as $p^\text{ld}$.

Expressions:

Subtract$p^\text{ld}$ and $q^\text{ld}$ terms and to the respective active and reactive power balance expressions created by the selected Network Formulations

Constraints:

\[\begin{aligned} & p_t^\text{ld} \le u_t^\text{ld} \cdot \text{ActivePowerTimeSeriesParameter}_t, \quad \forall t \in \{1,\dots, T\} \\ & q_t^\text{re} = \text{pf} \cdot p_t^\text{re}, \quad \forall t \in \{1,\dots, T\} \end{aligned}\]

on which $\text{pf} = \sin(\arctan(Q^\text{ld,max}/P^\text{ld,max}))$.

PowerLoadDispatch

PowerSimulations.PowerLoadDispatch — Type

Formulation type to enable (continuous) load interruption dispatch

Variables:

ActivePowerVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_active_power(device)

Symbol: $p^\text{ld}$

ReactivePowerVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_reactive_power(device)

Symbol: $q^\text{ld}$

Static Parameters:

$P^\text{ld,max}$ = PowerSystems.get_max_active_power(device)

$Q^\text{ld,max}$ = PowerSystems.get_max_reactive_power(device)

Time Series Parameters:

Parameter	Default Time Series Name

ReactivePowerTimeSeriesParameter	max_active_power

ActivePowerTimeSeriesParameter	max_active_power

Objective:

Creates an objective function term based on the FunctionData Options where the quantity term is defined as $p^\text{ld}$.

Expressions:

Subtract$p^\text{ld}$ and $q^\text{ld}$ terms and to the respective active and reactive power balance expressions created by the selected Network Formulations

Constraints:

\[\begin{aligned} & p_t^\text{ld} \le \text{ActivePowerTimeSeriesParameter}_t, \quad \forall t \in \{1,\dots, T\}\\ & q_t^\text{ld} = \text{pf} \cdot p_t^\text{ld}, \quad \forall t \in \{1,\dots, T\}\\ \end{aligned}\]

on which $\text{pf} = \sin(\arctan(Q^\text{ld,max}/P^\text{ld,max}))$.

Valid configurations

Valid DeviceModels for subtypes of ElectricLoad include the following:

Valid DeviceModel	Device Type	Formulation

DeviceModel(InterruptiblePowerLoad, PowerLoadDispatch)	InterruptiblePowerLoad	PowerLoadDispatch

DeviceModel(InterruptibleStandardLoad, PowerLoadDispatch)	InterruptibleStandardLoad	PowerLoadDispatch

DeviceModel(ShiftablePowerLoad, PowerLoadDispatch)	ShiftablePowerLoad	PowerLoadDispatch

DeviceModel(ExponentialLoad, PowerLoadDispatch)	ExponentialLoad	PowerLoadDispatch

DeviceModel(MotorLoad, PowerLoadDispatch)	MotorLoad	PowerLoadDispatch

DeviceModel(PowerLoad, PowerLoadDispatch)	PowerLoad	PowerLoadDispatch

DeviceModel(StandardLoad, PowerLoadDispatch)	StandardLoad	PowerLoadDispatch

DeviceModel(InterruptiblePowerLoad, PowerLoadInterruption)	InterruptiblePowerLoad	PowerLoadInterruption

DeviceModel(InterruptibleStandardLoad, PowerLoadInterruption)	InterruptibleStandardLoad	PowerLoadInterruption

DeviceModel(ShiftablePowerLoad, PowerLoadInterruption)	ShiftablePowerLoad	PowerLoadInterruption

DeviceModel(ExponentialLoad, PowerLoadInterruption)	ExponentialLoad	PowerLoadInterruption

DeviceModel(MotorLoad, PowerLoadInterruption)	MotorLoad	PowerLoadInterruption

DeviceModel(PowerLoad, PowerLoadInterruption)	PowerLoad	PowerLoadInterruption

DeviceModel(StandardLoad, PowerLoadInterruption)	StandardLoad	PowerLoadInterruption

DeviceModel(FixedAdmittance, StaticPowerLoad)	FixedAdmittance	StaticPowerLoad

DeviceModel(InterruptiblePowerLoad, StaticPowerLoad)	InterruptiblePowerLoad	StaticPowerLoad

DeviceModel(InterruptibleStandardLoad, StaticPowerLoad)	InterruptibleStandardLoad	StaticPowerLoad

DeviceModel(ShiftablePowerLoad, StaticPowerLoad)	ShiftablePowerLoad	StaticPowerLoad

DeviceModel(ExponentialLoad, StaticPowerLoad)	ExponentialLoad	StaticPowerLoad

DeviceModel(MotorLoad, StaticPowerLoad)	MotorLoad	StaticPowerLoad

DeviceModel(PowerLoad, StaticPowerLoad)	PowerLoad	StaticPowerLoad

DeviceModel(StandardLoad, StaticPowerLoad)	StandardLoad	StaticPowerLoad

DeviceModel(SwitchedAdmittance, StaticPowerLoad)	SwitchedAdmittance	StaticPowerLoad