PowerSystems.RenewableGen Formulations

Renewable generation formulations define the optimization models that describe renewable units mathematical model in different operational settings, such as economic dispatch and unit commitment.

Note

The use of reactive power variables and constraints will depend on the network model used, i.e., whether it uses (or does not use) reactive power. If the network model is purely active power-based, reactive power variables and related constraints are not created.

Note

Reserve variables for services are not included in the formulation, albeit their inclusion change the variables, expressions, constraints and objective functions created. A detailed description of the implications in the optimization models is described in the Service formulation section.

Table of contents

RenewableFullDispatch

RenewableConstantPowerFactor

Valid configurations

RenewableFullDispatch

PowerSimulations.RenewableFullDispatch — Type

Formulation type to add injection variables constrained by a maximum injection time series for RenewableGen

Variables:

ActivePowerVariable:

Bounds: [0.0, ]

Symbol: $p^\text{re}$

ReactivePowerVariable:

Bounds: [0.0, ]

Symbol: $q^\text{re}$

Static Parameters:

$P^\text{re,min}$ = PowerSystems.get_active_power_limits(device).min

$Q^\text{re,min}$ = PowerSystems.get_reactive_power_limits(device).min

$Q^\text{re,max}$ = PowerSystems.get_reactive_power_limits(device).max

Time Series Parameters:

Uses the max_active_power timeseries parameter to limit the available active power at each time-step.

Parameter	Default Time Series Name

ActivePowerTimeSeriesParameter	max_active_power

ReactivePowerTimeSeriesParameter	max_active_power

Objective:

Creates an objective function term based on the FunctionData Options where the quantity term is defined as $- p^\text{re}$ to incentivize generation from RenewableGen devices.

Expressions:

Adds $p^\text{re}$ and $q^\text{re}$ terms to the respective active and reactive power balance expressions created by the selected Network Formulations.

Constraints:

\[\begin{aligned} & P^\text{re,min} \le p_t^\text{re} \le \text{ActivePowerTimeSeriesParameter}_t, \quad \forall t \in \{1,\dots, T\} \\ & Q^\text{re,min} \le q_t^\text{re} \le Q^\text{re,max}, \quad \forall t \in \{1,\dots, T\} \end{aligned}\]

RenewableConstantPowerFactor

PowerSimulations.RenewableConstantPowerFactor — Type

Formulation type to add real and reactive injection variables with constant power factor with maximum real power injections constrained by a time series for RenewableGen

Variables:

ActivePowerVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_active_power(device)

Symbol: $p^\text{re}$

ReactivePowerVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_reactive_power(device)

Symbol: $q^\text{re}$

Static Parameters:

$P^\text{re,min}$ = PowerSystems.get_active_power_limits(device).min

$Q^\text{re,min}$ = PowerSystems.get_reactive_power_limits(device).min

$Q^\text{re,max}$ = PowerSystems.get_reactive_power_limits(device).max

$\text{pf}$ = PowerSystems.get_power_factor(device)

Time Series Parameters:

Parameter	Default Time Series Name

ActivePowerTimeSeriesParameter	max_active_power

ReactivePowerTimeSeriesParameter	max_active_power

Objective:

Creates an objective function term based on the FunctionData Options where the quantity term is defined as $- p_t^\text{re}$ to incentivize generation from RenewableGen devices.

Expressions:

Adds $p^\text{re}$ and $q^\text{re}$ terms to the respective active and reactive power balance expressions created by the selected Network Formulations

Constraints:

\[\begin{aligned} & P^\text{re,min} \le p_t^\text{re} \le \text{ActivePowerTimeSeriesParameter}_t, \quad \forall t \in \{1,\dots, T\} \\ & q_t^\text{re} = \text{pf} \cdot p_t^\text{re}, \quad \forall t \in \{1,\dots, T\} \end{aligned}\]

Valid configurations

Valid DeviceModels for subtypes of RenewableGen include the following:

Valid DeviceModel	Device Type	Formulation

DeviceModel(RenewableDispatch, FixedOutput)	RenewableDispatch	FixedOutput

DeviceModel(RenewableNonDispatch, FixedOutput)	RenewableNonDispatch	FixedOutput

DeviceModel(RenewableDispatch, RenewableConstantPowerFactor)	RenewableDispatch	RenewableConstantPowerFactor

DeviceModel(RenewableNonDispatch, RenewableConstantPowerFactor)	RenewableNonDispatch	RenewableConstantPowerFactor

DeviceModel(RenewableDispatch, RenewableFullDispatch)	RenewableDispatch	RenewableFullDispatch

DeviceModel(RenewableNonDispatch, RenewableFullDispatch)	RenewableNonDispatch	RenewableFullDispatch